### 2024-12-15 #222 @lucasimi
tda-mapper submission
Submitting Author: (@lucasimi)
Package Name: tda-mapper
One-Line Description of Package: A Python library implementing the Mapper algorithm for Topological Data Analysis.
Repository Link: https://github.com/lucasimi/tda-mapper-python
Version submitted: v0.9.0
EiC: TBD
Version accepted: TBD

---

- Include a brief paragraph describing what your package does: `tda-mapper` is a Python library that provides an efficient implementation of the Mapper algorithm, a powerful tool for topological data analysis. The algorithm transforms high-dimensional and complex datasets into graph representations, that are visualized through interactive plots, allowing users to explore hidden patterns, relationships, and structures within the data.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

  This library falls under the categories of "data processing/munging" and "data visualization" because it uses the Mapper algorithm to transform complex datasets into network representations, enabling users to process, analyze, and visually explore underlying structures and relationships.

  This package is aimed at researchers and data scientists engaged in exploratory data analysis. The Mapper algorithm is particularly useful in the early stages of data exploration, helping to uncover patterns and structures that guide further, more detailed analysis. It has been successfully applied in diverse fields, including social sciences, biology, and machine learning, to gain insights into complex datasets.

  Several Python packages, such as [GUDHI](https://github.com/GUDHI), [giotto-tda](https://github.com/giotto-ai/giotto-tda), and [Kepler Mapper](https://github.com/scikit-tda/kepler-mapper), offer implementations of the Mapper algorithm. However, tda-mapper differs from them by prioritizing performance and scalability in higher-dimensional spaces. Specifically, it efficiently computes Mapper on high-dimensional "lenses" that are computationally challenging for traditional methods. This approach not only enables the handling of larger and more complex datasets but also results in Mapper graphs that are easier to interpret and navigate. The approach used by tda-mapper scales better with dimension, making it faster and more responsive for interactive explorations compared to conventional techniques.

  This is the link to the pre-submission enquiry: https://github.com/pyOpenSci/software-submission/issues/219. I should also report that the methodology that this package is based on is explained more deeply in the [preprint](https://doi.org/10.5281/zenodo.10659651) and is currently under review for publication in a peer-review scientific journal.
