### 2024-12-25 68bbb29
Improved inset plot
