### 2025-01-11 9714369
Add flexible noise point handling in clustering
Add NoiseHandlingClustering wrapper class to provide control over how 
noise points (labeled as -1) are handled during clustering. The wrapper 
supports three modes:
- 'singleton': Convert each noise point to its own cluster
- 'drop': Keep noise points labeled as -1
- 'group': Group all noise points into a single cluster

Update mapper_connected_components documentation to clarify noise point 
behavior and add comprehensive unit tests for the new functionality. 
This change enables more flexible handling of outliers and noise points 
in different
[…]
