### 2025-01-11 39a91d5
Here's a git commit message based on the diff:
Optimize noise handling in clustering implementation

- Improve NoiseHandlingClustering performance with numpy operations
- Add validation for noise_handling parameter values
- Add detailed performance implications documentation
- Enhance code readability with clearer variable names
- Add debug output and more thorough tests for noise handling
- Fix potential edge case when all points are noise
- Use array copying to preserve original cluster labels

The changes focus on making the noise handling more robust and
efficient while maintaining the sam
[…]
