### 2025-01-21 d2dc87e
Added references to paper
