### 2025-01-21 6ee2583
Fixed scikit-learn dev dependency version
