+++ src/tdamapper/_common.py
@@ -1,16 +1,12 @@
 """
 This module provides common functionalities for internal use.
 """
+
 import warnings
 
 import numpy as np
 
-
-warnings.filterwarnings(
-    'default',
-    category=DeprecationWarning,
-    module=r'^tdamapper\.'
-)
+warnings.filterwarnings("default", category=DeprecationWarning, module=r"^tdamapper\.")
 
 
 def deprecated(msg):
@@ -18,7 +14,9 @@ def deprecated(msg):
         def wrapper(*args, **kwargs):
             warnings.warn(msg, DeprecationWarning, stacklevel=2)
             return func(*args, **kwargs)
+
         return wrapper
+
     return deprecated_func
 
 
@@ -30,34 +28,28 @@ class EstimatorMixin:
 
     def __is_sparse(self, X):
         # simple alternative use scipy.sparse.issparse
-        return hasattr(X, 'toarray')
+        return hasattr(X, "toarray")
 
     def _validate_X_y(self, X, y):
         if self.__is_sparse(X):
-            raise ValueError('Sparse data not supported.')
+            raise ValueError("Sparse data not supported.")
 
         X = np.asarray(X)
         y = np.asarray(y)
 
         if X.size == 0:
-            msg = (
-                f'0 feature(s) (shape={X.shape}) while a minimum of 1 is '
-                'required.'
-            )
+            msg = f"0 feature(s) (shape={X.shape}) while a minimum of 1 is " "required."
             raise ValueError(msg)
 
         if y.size == 0:
-            msg = (
-                f'0 feature(s) (shape={y.shape}) while a minimum of 1 is '
-                'required.'
-            )
+            msg = f"0 feature(s) (shape={y.shape}) while a minimum of 1 is " "required."
             raise ValueError(msg)
 
         if X.ndim == 1:
-            raise ValueError('1d-arrays not supported.')
+            raise ValueError("1d-arrays not supported.")
 
         if np.iscomplexobj(X) or np.iscomplexobj(y):
-            raise ValueError('Complex data not supported.')
+            raise ValueError("Complex data not supported.")
 
         if X.dtype == np.object_:
             X = np.array(X, dtype=float)
@@ -65,9 +57,13 @@ class EstimatorMixin:
         if y.dtype == np.object_:
             y = np.array(y, dtype=float)
 
-        if np.isnan(X).any() or np.isinf(X).any() or \
-                np.isnan(y).any() or np.isinf(y).any():
-            raise ValueError('NaNs or infinite values not supported.')
+        if (
+            np.isnan(X).any()
+            or np.isinf(X).any()
+            or np.isnan(y).any()
+            or np.isinf(y).any()
+        ):
+            raise ValueError("NaNs or infinite values not supported.")
 
         return X, y
 
@@ -82,12 +78,12 @@ class ParamsMixin:
     """
 
     def __is_param_public(self, k):
-        return (not k.startswith('_')) and (not k.endswith('_'))
+        return (not k.startswith("_")) and (not k.endswith("_"))
 
     def __split_param(self, k):
-        k_split = k.split('__')
+        k_split = k.split("__")
         outer = k_split[0]
-        inner = '__'.join(k_split[1:])
+        inner = "__".join(k_split[1:])
         return outer, inner
 
     def get_params(self, deep=True):
@@ -101,9 +97,9 @@ class ParamsMixin:
         for k, v in self.__dict__.items():
             if self.__is_param_public(k):
                 params[k] = v
-                if hasattr(v, 'get_params') and deep:
+                if hasattr(v, "get_params") and deep:
                     for _k, _v in v.get_params().items():
-                        params[f'{k}__{_k}'] = _v
[…]
