### 2025-04-17 4221b38
Converted notebooks to python files. Added steps for conversion and rendering.
