+++ docs/requirements.txt
@@ -1,7 +1,8 @@
 sphinx_rtd_theme
 nbsphinx
-ipykernel
 jupyter
 pandoc
 jupytext
+pandas
+scikit-learn<1.6.0
 tda-mapper
