### 2025-04-17 8cba665
Improved plots
