### 2025-04-18 e32a307
Improved notebooks and README
