+++ docs/source/notebooks/digits_online.py
@@ -17,74 +17,78 @@
 
 # %%
 import numpy as np
-
-from sklearn.datasets import load_digits
 from sklearn.cluster import AgglomerativeClustering
+from sklearn.datasets import load_digits
 from sklearn.decomposition import PCA
 
-from tdamapper.learn import MapperAlgorithm
-from tdamapper.cover import CubicalCover
 from tdamapper.clustering import FailSafeClustering
+from tdamapper.cover import CubicalCover
+from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
 
+# We load a labelled dataset
+X, labels = load_digits(return_X_y=True)
 
-X, y = load_digits(return_X_y=True)                 # We load a labelled dataset
-lens = PCA(2, random_state=42).fit_transform(X)     # We compute the lens values
+# Apply PCA as lens
+y = PCA(2, random_state=42).fit_transform(X)
 
 # %% [markdown]
 # ### Build Mapper graph
 
 # %%
-mapper_algo = MapperAlgorithm(
-    cover=CubicalCover(
-        n_intervals=10,
-        overlap_frac=0.65
-    ),
+algo = MapperAlgorithm(
+    cover=CubicalCover(n_intervals=10, overlap_frac=0.5),
     clustering=AgglomerativeClustering(10),
-    verbose=False
+    verbose=False,
 )
 
-mapper_graph = mapper_algo.fit_transform(X, lens)
+graph = algo.fit_transform(X, y)
 
 # %% [markdown]
 # ### Plot Mapper graph with mean
 
 # %%
-mapper_plot = MapperPlot(
-    mapper_graph,
-    dim=2,
-    iterations=400,
-    seed=42
-)
+plot = MapperPlot(graph, dim=3, iterations=400, seed=42)
 
-fig = mapper_plot.plot_plotly(
-    colors=y,                        # We color according to digit values
-    cmap='jet',                      # Jet colormap, used for classes
-    agg=np.nanmean,                  # We aggregate on graph nodes according to mean
-    title='digit (mean)',
+fig = plot.plot_plotly(
+    colors=labels,  # We color according to digit values
+    cmap="jet",  # Jet colormap, used for classes
+    agg=np.nanmean,  # We aggregate on graph nodes according to mean
+    title="digit (mean)",
     width=600,
-    height=600
+    height=600,
 )
 
-fig.show(
-    renderer='notebook_connected',
-    config={'scrollZoom': True}
-)
+fig.show(renderer="notebook_connected", config={"scrollZoom": True})
 
 # %% [markdown]
 # ### Plot Mapper graph with standard deviation
 
 # %%
-fig = mapper_plot.plot_plotly(
-    colors=y,                        
-    cmap='viridis',                  # Viridis colormap, used for ranges
-    agg=np.nanstd,                   # We aggregate on graph nodes according to std
-    title='digit (std)',
+fig = plot.plot_plotly(
+    colors=labels,
+    cmap="viridis",  # Viridis colormap, used for ranges
+    agg=np.nanstd,  # We aggregate on graph nodes according to std
+    title="digit (std)",
     width=600,
-    height=600
+    height=600,
 )
 
-fig.show(
-    renderer='notebook_connected',
-    config={'scrollZoom': True}
-)
+fig.show(renderer="notebook_connected", config={"scrollZoom": True})
+
+# %% [Markdown]
+# ### Inspect interesting nodes
+
+# %%
+from matplotlib import pyplot as plt
+
+# By interacting with the plot we see that node 140 is joining the cluster of
[…]
