+++ README.md
@@ -92,7 +92,9 @@ from tdamapper.plot import MapperPlot
 
 # Generate toy dataset
 X, labels = make_circles(n_samples=5000, noise=0.05, factor=0.3, random_state=42)
-plt.scatter(X[:,0], X[:,1], c=labels, cmap='jet', s=0.25)
+plt.figure(figsize=(5, 5))
+plt.scatter(X[:,0], X[:,1], c=labels, s=0.25, cmap="jet")
+plt.axis("off")
 plt.show()
 
 # Apply PCA as lens
@@ -105,7 +107,7 @@ graph = MapperAlgorithm(cover, clust).fit_transform(X, y)
 
 # Visualize the Mapper graph
 fig = MapperPlot(graph, dim=2, seed=42, iterations=60).plot_plotly(colors=labels)
-fig.show(config={'scrollZoom': True})
+fig.show(config={"scrollZoom": True})
 ```
 
 | Original Dataset | Mapper Graph |
