### 2025-04-18 854b64f
Fixed cell rendering
