### 2025-04-18 6cf358f
Added node_size param
