### 2025-04-20 dc4c7e6
Improved package settings
