+++ benchmarks/plot.py
@@ -1,81 +1,70 @@
 import math
 
-import pandas as pd
 import matplotlib.pyplot as plt
-from matplotlib.ticker import MaxNLocator, LogLocator
-
+import pandas as pd
+from matplotlib.ticker import LogLocator, MaxNLocator
 
 alpha = 0.75
 
 
-df_colors = pd.DataFrame({
-    'bench': ['tda-mapper', 'giotto-tda', 'kepler-mapper'],
-    'color': ['tab:blue', 'tab:orange', 'tab:green'],
-})
+df_colors = pd.DataFrame(
+    {
+        "bench": ["tda-mapper", "giotto-tda", "kepler-mapper"],
+        "color": ["tab:blue", "tab:orange", "tab:green"],
+    }
+)
 
 
-df_styles = pd.DataFrame({
-    'bench': ['tda-mapper', 'giotto-tda', 'kepler-mapper'],
-    'style': ['-', '--', '-.'],
-})
+df_styles = pd.DataFrame(
+    {
+        "bench": ["tda-mapper", "giotto-tda", "kepler-mapper"],
+        "style": ["-", "--", "-."],
+    }
+)
 
 
-df_markers = pd.DataFrame({
-    'bench': ['tda-mapper', 'giotto-tda', 'kepler-mapper'],
-    'marker': ['D', 's', 'o'],
-})
+df_markers = pd.DataFrame(
+    {
+        "bench": ["tda-mapper", "giotto-tda", "kepler-mapper"],
+        "marker": ["D", "s", "o"],
+    }
+)
 
 
-df_zorder = pd.DataFrame({
-    'bench': ['tda-mapper', 'giotto-tda', 'kepler-mapper'],
-    'zorder': [3, 1, 2],
-})
+df_zorder = pd.DataFrame(
+    {
+        "bench": ["tda-mapper", "giotto-tda", "kepler-mapper"],
+        "zorder": [3, 1, 2],
+    }
+)
 
 
-df_titles = pd.DataFrame({
-    'dataset': ['line', 'digits', 'mnist', 'cifar10', 'fashion_mnist'],
-    'title': ['Line', 'Digits', 'MNIST', 'Cifar-10', 'Fashion-MNIST'],
-})
+df_titles = pd.DataFrame(
+    {
+        "dataset": ["line", "digits", "mnist", "cifar10", "fashion_mnist"],
+        "title": ["Line", "Digits", "MNIST", "Cifar-10", "Fashion-MNIST"],
+    }
+)
 
 
 def load_benchmark(path):
     df = pd.read_csv(path)
-    df = pd.merge(
-        left=df,
-        right=df_colors,
-        on='bench'
-    )
-    df = pd.merge(
-        left=df,
-        right=df_styles,
-        on='bench'
-    )
-    df = pd.merge(
-        left=df,
-        right=df_markers,
-        on='bench'
-    )
-    df = pd.merge(
-        left=df,
-        right=df_zorder,
-        on='bench'
-    )
-    df = pd.merge(
-        left=df,
-        right=df_titles,
-        on='dataset'
-    )
-    df.sort_values(by=['dataset', 'bench', 'p', 'k'], inplace=True)
+    df = pd.merge(left=df, right=df_colors, on="bench")
+    df = pd.merge(left=df, right=df_styles, on="bench")
+    df = pd.merge(left=df, right=df_markers, on="bench")
+    df = pd.merge(left=df, right=df_zorder, on="bench")
+    df = pd.merge(left=df, right=df_titles, on="dataset")
+    df.sort_values(by=["dataset", "bench", "p", "k"], inplace=True)
     df.reset_index(drop=True, inplace=True)
     return df
 
 
[…]
