+++ tests/test_unit_heap.py
@@ -1,5 +1,5 @@
-import unittest
 import random
+import unittest
 
 from tdamapper.utils.heap import MaxHeap
 
