+++ tests/test_unit_plot.py
@@ -1,55 +1,44 @@
 import unittest
 
+import networkx as nx
 import numpy as np
 
-import networkx as nx
-
+from tdamapper.clustering import TrivialClustering
 from tdamapper.core import MapperAlgorithm
 from tdamapper.cover import BallCover
-from tdamapper.clustering import TrivialClustering
-from tdamapper.plot import MapperPlot
-from tdamapper.plot import MapperLayoutInteractive, MapperLayoutStatic
+from tdamapper.plot import MapperLayoutInteractive, MapperLayoutStatic, MapperPlot
 
 
 class TestMapperPlot(unittest.TestCase):
 
     def test_two_connected_clusters(self):
         data = [
-            np.array([0.0, 1.0]), np.array([1.0, 0.0]),
-            np.array([0.0, 0.0]), np.array([1.0, 1.0])
+            np.array([0.0, 1.0]),
+            np.array([1.0, 0.0]),
+            np.array([0.0, 0.0]),
+            np.array([1.0, 1.0]),
         ]
         mp = MapperAlgorithm(
-            cover=BallCover(1.1, metric='euclidean'),
-            clustering=TrivialClustering()
+            cover=BallCover(1.1, metric="euclidean"), clustering=TrivialClustering()
         )
         g = mp.fit_transform(data, data)
-        mp_plot1 = MapperPlot(
-            g,
-            dim=2,
-            seed=123,
-            iterations=10
-        )
+        mp_plot1 = MapperPlot(g, dim=2, seed=123, iterations=10)
         fig1 = mp_plot1.plot_plotly(
             colors=data,
             agg=np.nanmax,
             width=200,
             height=200,
-            title='example',
-            cmap='jet'
+            title="example",
+            cmap="jet",
         )
-        mp_plot2 = MapperPlot(
-            g,
-            dim=3,
-            seed=123,
-            iterations=10
-        )
+        mp_plot2 = MapperPlot(g, dim=3, seed=123, iterations=10)
         fig2 = mp_plot2.plot_plotly(
             colors=data,
             agg=np.nanmax,
             width=200,
             height=200,
-            title='example',
-            cmap='jet'
+            title="example",
+            cmap="jet",
         )
         fig3 = mp_plot2.plot_plotly_update(
             fig2,
@@ -57,23 +46,16 @@ class TestMapperPlot(unittest.TestCase):
             agg=np.nanmin,
             width=300,
             height=300,
-            title='example-updated',
-            cmap='viridis'
+            title="example-updated",
+            cmap="viridis",
         )
-        mp_plot3 = MapperPlot(
-            g,
-            dim=2
-        )
-        fig4 = mp_plot3.plot_matplotlib(
-            width=300,
-            height=300,
-            colors=data
-        )
+        mp_plot3 = MapperPlot(g, dim=2)
+        fig4 = mp_plot3.plot_matplotlib(width=300, height=300, colors=data)
         mp_plot3.plot_pyvis(
             width=512,
             height=512,
             colors=data,
-            output_file='network.html',
+            output_file="network.html",
         )
 
     def test_empty_graph(self):
@@ -81,15 +63,17 @@ class TestMapperPlot(unittest.TestCase):
         mapper_plot = MapperPlot(empty_graph, dim=2)
         mapper_plot.plot_matplotlib(colors=[])
         mapper_plot.plot_plotly(colors=[])
-        mapper_plot.plot_pyvis(colors=[], output_file='tmp.html')
+        mapper_plot.plot_pyvis(colors=[], output_file="tmp.html")
 
     def test_two_connected_clusters_deprecated(self):
         data = [
-            np.array([0.0, 1.0]), np.array([1.0, 0.0]),
[…]
