+++ tests/test_unit_learn.py
@@ -1,18 +1,16 @@
-
 import unittest
 
-import numpy as np
 import networkx as nx
-
-from sklearn.utils.estimator_checks import check_estimator
+import numpy as np
 from sklearn.cluster import DBSCAN
+from sklearn.utils.estimator_checks import check_estimator
 
+from tdamapper.core import TrivialClustering, TrivialCover
+from tdamapper.cover import BallCover
 from tdamapper.learn import (
     MapperAlgorithm,
     MapperClustering,
 )
-from tdamapper.core import TrivialClustering, TrivialCover
-from tdamapper.cover import BallCover
 
 
 def euclidean(x, y):
