+++ docs/source/notebooks/digits.py
@@ -1,0 +1,197 @@
+# ---
+# jupyter:
+#   jupytext:
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.17.0
+#   kernelspec:
+#     display_name: Python 3 (ipykernel)
+#     language: python
+#     name: python3
+# ---
+
+# %% [markdown]
+# # Example 2: Confusion Analysis using Mapper Algorithm
+# In this notebook, we analyze a dataset of handwritten digits using the
+# **Mapper algorithm**, a powerful tool from topological data analysis (TDA).
+# Mapper helps us uncover regions in the dataset where samples are hard to
+# classify. By applying Mapper to the digit dataset, we generate a **graph**
+# where each node represents a group of similar samples. We color the nodes
+# based on the **entropy** of their digit labels, allowing us to identify areas
+# of high ambiguity. These high-entropy nodes typically contain images that are
+# visually similar but belong to different digit classes, making them more
+# challenging to classify correctly. The goal of this analysis is to
+# demonstrate how Mapper can provide insights into classification problems by
+# revealing these ambiguous regions and offering an intuitive way to explore them.
+
+
+# %% [markdown]
+# ### Mapper pipeline
+
+# %% [markdown]
+# To begin, we load the **Digits dataset**, which consists of 8x8 pixel images
+# of handwritten digits. We then apply **Principal Component Analysis (PCA)**
+# to reduce the dimensionality of the dataset from 64 features to just 2. These
+# two principal components will serve as the input for the Mapper algorithm,
+# providing a lower-dimensional "lens" through which we will analyze the
+# structure of the data.
+
+# %%
+import numpy as np
+from sklearn.cluster import AgglomerativeClustering
+from sklearn.datasets import load_digits
+from sklearn.decomposition import PCA
+
+from tdamapper.clustering import FailSafeClustering
+from tdamapper.cover import CubicalCover
+from tdamapper.learn import MapperAlgorithm
+from tdamapper.plot import MapperPlot
+
+X, labels = load_digits(return_X_y=True)
+y = PCA(2, random_state=42).fit_transform(X)
+
+# %% [markdown]
+# The **Mapper algorithm** relies on two key parameters: **cover** and
+# **clustering**. The **cover** defines how the data is partitioned into
+# intervals, while the **clustering** method groups the data points into
+# clusters based on their proximity in the reduced space. Choosing the right
+# settings for these parameters is crucial, as they directly influence the
+# structure of the Mapper graph. In this notebook, we use a **CubicalCover**
+# with 10 intervals and 50% overlap between intervals, along with
+# **AgglomerativeClustering** to form 10 clusters. However, finding the best
+# parameters often requires experimentation based on the specific dataset and
+# the problem at hand.
+
+
+# %%
+mapper = MapperAlgorithm(
+    cover=CubicalCover(n_intervals=10, overlap_frac=0.5),
+    clustering=AgglomerativeClustering(10),
+    verbose=False,
+)
+
+graph = mapper.fit_transform(X, y)
+
+# %% [markdown]
+# ### Visualization
+
+# %% [markdown]
+
+# To explore the results of the Mapper algorithm, we visualize the Mapper
+# graph, where each node represents a cluster of similar images. We color the
+# nodes based on the **mode** of the digit labels in each cluster. The **mode**
+# gives us a sense of the most common digit in each cluster. As shown in the
+# plot, the nodes are well-clustered by color, indicating that the digit labels
+# align well with the topological structure of the dataset. This suggests that
[…]
