+++ app/requirements.txt
@@ -1,7 +1,7 @@
 streamlit>=1.40.0,<2.0.0
 numpy>=1.25.2,<2.0.0
 scikit-learn>=1.5.0,<1.6.0
-umap-learn>=0.5.7,>0.6.0
+umap-learn>=0.5.7,<0.6.0
 pandas>=2.1.0,<3.0.0
 tda-mapper>=0.9.0,<0.10.0
 plotly>=6.0.0,<7.0.0
