### 2025-04-21 71eaa8a
Improved CI, updated info. Minor fixes to tests
