+++ .github/workflows/test-bench.yml
@@ -3,22 +3,38 @@ name: test-bench
 on: 
   push:
     branches:
-      - develop
-      - main      
+      - main
+  pull_request:
+    branches:
+      - main 
 
 jobs:
   test-bench-job:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
     steps:
       - name: Check out repository code
         uses: actions/checkout@v4
+
+      - name: Cache pip
+        uses: actions/cache@v4
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
+
       - name: Install Python
         uses: actions/setup-python@v5
         with:
-          python-version: '3.10' 
+          python-version: ${{ matrix.python-version }}
+
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
+
       - name: Run benchmarks
         run: | 
           python -m unittest discover -s tests -p 'test_bench_*.py'
