+++ .github/workflows/test-unit.yml
@@ -2,27 +2,44 @@ name: test-unit
 
 on: 
   push:
-    branches:
-      - develop
-      - main      
+    branches: 
+      - '**'
+  pull_request:
+    branches: 
+      - '**'
 
 jobs:
   test-unit-job:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
     steps:
       - name: Check out repository code
         uses: actions/checkout@v4
+
+      - name: Cache pip
+        uses: actions/cache@v4
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
+
       - name: Install Python
         uses: actions/setup-python@v5
         with:
-          python-version: '3.10' 
+          python-version: ${{ matrix.python-version }}
+      
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
+
       - name: Run tests and code coverage
         run: | 
           coverage run --source=src -m unittest discover -s tests -p 'test_unit_*.py'
           coverage report -m
+
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v4
         env:
