### 2025-04-21 865582f
Improved CI with EOL python versions
