+++ .github/workflows/test-unit.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13", "3.14"]
     steps:
       - name: Check out repository code
         uses: actions/checkout@v4
@@ -34,11 +34,21 @@ jobs:
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
+        continue-on-error: ${{ 
+          matrix.python-version == '3.6' || 
+          matrix.python-version == '3.7' || 
+          matrix.python-version == '3.8' 
+        }} # EOL python versions are allowed to fail
 
       - name: Run tests and code coverage
         run: | 
           coverage run --source=src -m unittest discover -s tests -p 'test_unit_*.py'
           coverage report -m
+        continue-on-error: ${{ 
+          matrix.python-version == '3.6' || 
+          matrix.python-version == '3.7' || 
+          matrix.python-version == '3.8' 
+        }} # EOL python versions are allowed to fail
 
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v4
