+++ .github/workflows/test-bench.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13", "3.14"]
     steps:
       - name: Check out repository code
         uses: actions/checkout@v4
@@ -34,7 +34,17 @@ jobs:
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
+        continue-on-error: ${{ 
+          matrix.python-version == '3.6' || 
+          matrix.python-version == '3.7' || 
+          matrix.python-version == '3.8' 
+        }} # EOL python versions are allowed to fail
 
       - name: Run benchmarks
         run: | 
           python -m unittest discover -s tests -p 'test_bench_*.py'
+        continue-on-error: ${{ 
+          matrix.python-version == '3.6' || 
+          matrix.python-version == '3.7' || 
+          matrix.python-version == '3.8' 
+        }} # EOL python versions are allowed to fail
