+++ .github/workflows/test-unit.yml
@@ -30,7 +30,7 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-      
+
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
