+++ .github/workflows/test-bench.yml
@@ -34,17 +34,19 @@ jobs:
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
-        continue-on-error: ${{ 
-          matrix.python-version == '3.6' || 
-          matrix.python-version == '3.7' || 
-          matrix.python-version == '3.8' 
-        }} # EOL python versions are allowed to fail
+        continue-on-error: | 
+          ${{ 
+            matrix.python-version == '3.6' || 
+            matrix.python-version == '3.7' || 
+            matrix.python-version == '3.8' 
+          }} # EOL python versions are allowed to fail
 
       - name: Run benchmarks
         run: | 
           python -m unittest discover -s tests -p 'test_bench_*.py'
-        continue-on-error: ${{ 
-          matrix.python-version == '3.6' || 
-          matrix.python-version == '3.7' || 
-          matrix.python-version == '3.8' 
-        }} # EOL python versions are allowed to fail
+        continue-on-error: |
+          ${{ 
+            matrix.python-version == '3.6' || 
+            matrix.python-version == '3.7' || 
+            matrix.python-version == '3.8' 
+          }} # EOL python versions are allowed to fail
