+++ .github/workflows/test-unit.yml
@@ -34,21 +34,23 @@ jobs:
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
-        continue-on-error: ${{ 
-          matrix.python-version == '3.6' || 
-          matrix.python-version == '3.7' || 
-          matrix.python-version == '3.8' 
-        }} # EOL python versions are allowed to fail
+        continue-on-error: | 
+          ${{ 
+            matrix.python-version == '3.6' || 
+            matrix.python-version == '3.7' || 
+            matrix.python-version == '3.8' 
+          }} # EOL python versions are allowed to fail
 
       - name: Run tests and code coverage
         run: | 
           coverage run --source=src -m unittest discover -s tests -p 'test_unit_*.py'
           coverage report -m
-        continue-on-error: ${{ 
-          matrix.python-version == '3.6' || 
-          matrix.python-version == '3.7' || 
-          matrix.python-version == '3.8' 
-        }} # EOL python versions are allowed to fail
+        continue-on-error: |
+          ${{ 
+            matrix.python-version == '3.6' || 
+            matrix.python-version == '3.7' || 
+            matrix.python-version == '3.8' 
+          }} # EOL python versions are allowed to fail
 
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v4
