+++ .github/workflows/test-bench.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   test-bench-job:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
     strategy:
       matrix:
         python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
