+++ .github/workflows/test-unit.yml
@@ -30,6 +30,12 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
+        continue-on-error: | 
+          ${{ 
+            matrix.python-version == '3.6' || 
+            matrix.python-version == '3.7' || 
+            matrix.python-version == '3.8' 
+          }} # EOL python versions are allowed to fail
 
       - name: Install dependencies
         run: | 
