+++ .github/workflows/test-unit.yml
@@ -13,6 +13,7 @@ jobs:
     runs-on: ubuntu-22.04
     strategy:
       matrix:
+        fail-fast: false
         python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
     steps:
       - name: Check out repository code
