+++ .github/workflows/test-bench.yml
@@ -31,29 +31,26 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-        continue-on-error: | 
-          ${{ 
+        continue-on-error: ${{ 
             matrix.python-version == '3.6' || 
             matrix.python-version == '3.7' || 
-            matrix.python-version == '3.8' 
+            matrix.python-version == '3.8'
           }}
 
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
-        continue-on-error: | 
-          ${{ 
+        continue-on-error: ${{ 
             matrix.python-version == '3.6' || 
             matrix.python-version == '3.7' || 
-            matrix.python-version == '3.8' 
+            matrix.python-version == '3.8'
           }}
 
       - name: Run benchmarks
         run: | 
           python -m unittest discover -s tests -p 'test_bench_*.py'
-        continue-on-error: |
-          ${{ 
+        continue-on-error: ${{ 
             matrix.python-version == '3.6' || 
             matrix.python-version == '3.7' || 
-            matrix.python-version == '3.8' 
+            matrix.python-version == '3.8'
           }}
