+++ .github/workflows/test-bench.yml
@@ -12,7 +12,6 @@ jobs:
   test-bench-job:
     runs-on: ubuntu-22.04
     strategy:
-      fail-fast: false
       matrix:
         python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
     steps:
