+++ pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
     "matplotlib>=3.3.4,<4.0.0",
     "networkx>=2.5,<3.0.0",
     "igraph>=0.11.8,<0.12.0",
-    "numba>=0.54,<0.60",
+    "numba>=0.54,<0.62",
     "pyvis>=0.3.2,<0.4.0",
     "plotly>=6.0.1,<7.0.0",
     "joblib>=1.4.2,<2.0.0",
