### 2025-04-21 fcf5a60
Set allow-failure
