+++ .github/workflows/test-unit.yml
@@ -14,6 +14,9 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
+        allow-failure:
+        - ${{ contains(fromJson('["3.6","3.7","3.8"]'), matrix.python-version) }}
+
     steps:
       - name: Check out repository code
         uses: actions/checkout@v4
@@ -30,30 +33,18 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-        continue-on-error: ${{ 
-            matrix.python-version == '3.6' || 
-            matrix.python-version == '3.7' || 
-            matrix.python-version == '3.8'
-          }}
+        continue-on-error: ${{ matrix.allow-failure }}
 
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
-        continue-on-error: ${{ 
-            matrix.python-version == '3.6' || 
-            matrix.python-version == '3.7' || 
-            matrix.python-version == '3.8'
-          }}
+        continue-on-error: ${{ matrix.allow-failure }}
 
       - name: Run tests and code coverage
         run: | 
           coverage run --source=src -m unittest discover -s tests -p 'test_unit_*.py'
           coverage report -m
-        continue-on-error: ${{ 
-            matrix.python-version == '3.6' || 
-            matrix.python-version == '3.7' || 
-            matrix.python-version == '3.8' 
-          }}
+        continue-on-error: ${{ matrix.allow-failure }}
 
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v4
