+++ .github/workflows/test-bench.yml
@@ -14,8 +14,25 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
-        allow-failure:
-        - ${{ contains(fromJson('["3.6","3.7","3.8"]'), matrix.python-version) }}
+        include:
+          - python-version: 3.6
+            allow-failure: true
+          - python-version: 3.7
+            allow-failure: true
+          - python-version: 3.8
+            allow-failure: true
+          - python-version: 3.9
+            allow-failure: false
+          - python-version: 3.10
+            allow-failure: false
+          - python-version: 3.11
+            allow-failure: false
+          - python-version: 3.12
+            allow-failure: false
+          - python-version: 3.13
+            allow-failure: false
+    continue-on-error: ${{ matrix.allow-failure }}
+    name: Python ${{ matrix.python-version }} bench tests
 
     steps:
       - name: Check out repository code
@@ -33,14 +50,11 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-        continue-on-error: ${{ matrix.allow-failure }}
 
       - name: Install dependencies
         run: | 
           python -m pip install -e .[dev]
-        continue-on-error: ${{ matrix.allow-failure }}
 
       - name: Run benchmarks
         run: | 
           python -m unittest discover -s tests -p 'test_bench_*.py'
-        continue-on-error: ${{ matrix.allow-failure }}
