+++ pyproject.toml
@@ -20,6 +20,12 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Operating System :: OS Independent",
 ]
 keywords = ["tda", "mapper", "topology", "topological data analysis"]
@@ -33,7 +39,7 @@ dependencies = [
     "plotly>=6.0.1,<7.0.0",
     "joblib>=1.4.2,<2.0.0",
 ]
-requires-python = ">=3.6"
+requires-python = ">=3.8"
 
 [project.optional-dependencies]
 dev = [
