### 2025-04-21 c3a4389
Removed unsupported python versions
