+++ .github/workflows/test-unit.yml
@@ -13,14 +13,10 @@ jobs:
     runs-on: ubuntu-22.04
     strategy:
       matrix:
-        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
         include:
-          - python-version: 3.6
-            allow-failure: true
-          - python-version: 3.7
-            allow-failure: true
           - python-version: 3.8
-            allow-failure: true
+            allow-failure: false
           - python-version: 3.9
             allow-failure: false
           - python-version: 3.10
