+++ README.md
@@ -10,7 +10,13 @@
 
 # tda-mapper
 
-**tda-mapper** is a Python library built around the Mapper algorithm, a core technique in Topological Data Analysis (TDA) for extracting topological structure from complex data. Designed for computational efficiency and scalability, it leverages optimized spatial search methods to support high-dimensional datasets. The library is well-suited for integration into machine learning pipelines, unsupervised learning tasks, and exploratory data analysis.
+**tda-mapper** is a Python library built around the Mapper algorithm, a core
+technique in Topological Data Analysis (TDA) for extracting topological
+structure from complex data. Designed for computational efficiency and
+scalability, it leverages optimized spatial search methods to support
+high-dimensional datasets. The library is well-suited for integration into
+machine learning pipelines, unsupervised learning tasks, and exploratory data
+analysis.
 
 Further details in the
 [documentation](https://tda-mapper.readthedocs.io/en/main/)
@@ -21,23 +27,37 @@ and in the
 
 - **Efficient construction**
     
-    Leverages optimized spatial search techniques and parallelization to accelerate the construction of Mapper graphs, supporting the analysis of high-dimensional datasets.
+    Leverages optimized spatial search techniques and parallelization to
+    accelerate the construction of Mapper graphs, supporting the analysis of
+    high-dimensional datasets.
 
 - **Scikit-learn integration**
 
-    Provides custom estimators that are fully compatible with scikit-learn's API, enabling seamless integration into scikit-learn pipelines for tasks such as dimensionality reduction, clustering, and feature extraction.
+    Provides custom estimators that are fully compatible with scikit-learn's
+    API, enabling seamless integration into scikit-learn pipelines for tasks
+    such as dimensionality reduction, clustering, and feature extraction.
 
 - **Flexible visualization**
 
-    Multiple visualization backends supported (e.g., Plotly, Matplotlib) for generating high-quality Mapper graph representations with adjustable layouts and styling.
+    Multiple visualization backends supported (e.g., Plotly, Matplotlib) for
+    generating high-quality Mapper graph representations with adjustable 
+    layouts and styling.
 
 - **Interactive app**
 
-    Provides an interactive web-based interface (via Streamlit) for dynamic exploration of Mapper graph structures, offering real-time adjustments to parameters and visualizations.
+    Provides an interactive web-based interface (via Streamlit) for dynamic
+    exploration of Mapper graph structures, offering real-time adjustments to
+    parameters and visualizations.
 
 ## Background
 
-The Mapper algorithm extracts topological features from complex datasets, representing them as graphs that highlight clusters, transitions, and key structural patterns. These insights reveal hidden data relationships and are applicable across diverse fields, including social sciences, biology, and machine learning. For an in-depth overview of Mapper, including its mathematical foundations and practical applications, read [the original paper](https://research.math.osu.edu/tgda/mapperPBG.pdf).
+The Mapper algorithm extracts topological features from complex datasets,
+representing them as graphs that highlight clusters, transitions, and key
+structural patterns. These insights reveal hidden data relationships and are
+applicable across diverse fields, including social sciences, biology, and
+machine learning. For an in-depth overview of Mapper, including its
+mathematical foundations and practical applications, read 
+[the original paper](https://research.math.osu.edu/tgda/mapperPBG.pdf).
 
 | Step 1 | Step 2 | Step 3 | Step 4 |
 | ------ | ------ | ------ | ------ |
@@ -46,7 +66,13 @@ The Mapper algorithm extracts topological features from complex datasets, repres
 
 ## Citations
 
[…]
