### 2025-05-14 cc33f65
Moved ui elements to plotly plot
