+++ app/streamlit_app.py
@@ -23,6 +23,7 @@ from sklearn.decomposition import PCA
 from umap import UMAP
 
 from tdamapper._plot_plotly import _marker_size
+from tdamapper.core import aggregate_graph
 from tdamapper.cover import BallCover, CubicalCover
 from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
@@ -677,7 +678,7 @@ def mapper_figure_section(df_X, df_y, mapper_plot):
         colors=colors,
         node_size=1.0,
         _agg=agg,
-        cmap="Viridis",
+        cmap="Jet",
         agg_name=agg_name,
         colors_feat=colors_feat,
     )
@@ -696,73 +697,76 @@ def mapper_figure_section(df_X, df_y, mapper_plot):
         scaleratio=1,
     )
 
+    menu_cmap = {}
     if dim == 2:
-        _set_cmap_buttons_2d(mapper_fig)
+        menu_cmap = _set_cmap_buttons_2d(mapper_fig)
+        menu_feature = _set_feature_select_button_2d(
+            mapper_plot, mapper_fig, df_X, df_y, agg
+        )
 
     elif dim == 3:
-        _set_cmap_buttons_3d(mapper_fig)
+        menu_cmap = _set_cmap_buttons_3d(mapper_fig)
+        menu_feature = _set_feature_select_button_3d(
+            mapper_plot, mapper_fig, df_X, df_y, agg
+        )
 
-    _set_node_size_slider(mapper_plot, mapper_fig)
+    slider_size = _set_node_size_slider(mapper_plot, mapper_fig)
+
+    mapper_fig.update_layout(
+        updatemenus=[menu_cmap, menu_feature],
+        sliders=[slider_size],
+        uirevision="constant",
+    )
 
     return mapper_fig
 
 
 def _set_cmap_buttons_2d(fig):
-    fig.update_layout(
-        updatemenus=[
+    return dict(
+        buttons=[
             dict(
-                buttons=[
-                    dict(
-                        label=cmap_name,
-                        method="restyle",
-                        args=[
-                            {
-                                "marker.colorscale": [cmap],
-                                "marker.line.colorscale": [cmap],
-                            },
-                            [0, 1],  # Trace indices
-                        ],
-                    )
-                    for cmap_name, cmap in V_CMAPS.items()
+                label=cmap_name,
+                method="restyle",
+                args=[
+                    {
+                        "marker.colorscale": [cmap],
+                        "marker.line.colorscale": [cmap],
+                    },
+                    [0, 1],  # Trace indices
                 ],
-                x=0,
-                xanchor="left",
-                y=0.75,
-                yanchor="bottom",
-                direction="down",
             )
+            for cmap_name, cmap in V_CMAPS.items()
         ],
-        uirevision="constant",
+        x=0,
+        xanchor="left",
+        y=0.75,
+        yanchor="bottom",
+        direction="down",
     )
 
 
 def _set_cmap_buttons_3d(fig):
-    fig.update_layout(
-        updatemenus=[
+    return dict(
+        buttons=[
             dict(
-                buttons=[
-                    dict(
-                        label=cmap_name,
-                        method="restyle",
-                        args=[
-                            {
-                                "marker.colorscale": [cmap, cmap],
-                                "marker.line.colorscale": [cmap, cmap],
-                                "line.colorscale": [cmap, cmap],
-                            },
-                            [0, 1],  # update both traces
-                        ],
-                    )
-                    for cmap_name, cmap in V_CMAPS.items()
+                label=cmap_name,
+                method="restyle",
+                args=[
+                    {
+                        "marker.colorscale": [cmap, cmap],
+                        "marker.line.colorscale": [cmap, cmap],
+                        "line.colorscale": [cmap, cmap],
+                    },
[…]
