### 2025-05-15 8e1fb7b
Fixed trace colors
