+++ tests/example.py
@@ -1,3 +1,4 @@
+import matplotlib.pyplot as plt
 import numpy as np
 from sklearn.cluster import DBSCAN
 from sklearn.datasets import make_circles
@@ -7,34 +8,21 @@ from tdamapper.cover import CubicalCover
 from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
 
-X, y = make_circles(  # load a labelled dataset
-    n_samples=5000, noise=0.05, factor=0.3, random_state=42
-)
-lens = PCA(2).fit_transform(X)
+# Generate toy dataset
+X, labels = make_circles(n_samples=5000, noise=0.05, factor=0.3, random_state=42)
+plt.figure(figsize=(5, 5))
+plt.scatter(X[:, 0], X[:, 1], c=labels, s=0.25, cmap="jet")
+plt.axis("off")
+plt.show()
 
-mapper_algo = MapperAlgorithm(
-    cover=CubicalCover(n_intervals=10, overlap_frac=0.3), clustering=DBSCAN()
-)
-mapper_graph = mapper_algo.fit_transform(X, lens)
+# Apply PCA as lens
+y = PCA(2, random_state=42).fit_transform(X)
 
-mapper_plot = MapperPlot(mapper_graph, dim=2, iterations=60, seed=42)
+# Mapper pipeline
+cover = CubicalCover(n_intervals=10, overlap_frac=0.3)
+clust = DBSCAN()
+graph = MapperAlgorithm(cover, clust).fit_transform(X, y)
 
-fig = mapper_plot.plot_plotly(
-    title="",
-    width=600,
-    height=600,
-    colors=y,  # color according to categorical values
-    cmap="jet",  # Jet colormap, for classes
-    agg=np.nanmean,  # aggregate on nodes according to mean
-)
-
-fig.show(config={"scrollZoom": True})
-
-mapper_plot.plot_plotly_update(
-    fig,  # reuse the plot with the same positions
-    colors=y,
-    cmap="viridis",  # viridis colormap, for ranges
-    agg=np.nanstd,  # aggregate on nodes according to std
-)
-
+# Visualize the Mapper graph
+fig = MapperPlot(graph, dim=2, seed=42, iterations=60).plot_plotly(colors=labels)
 fig.show(config={"scrollZoom": True})
