### 2025-05-16 79dcb41
Added dev dependency for CI
