### 2025-05-16 7f80242
Removed calls to plt.show
