### 2025-05-16 050b074
Added color column and cmap selection
