+++ src/tdamapper/_plot_plotly.py
@@ -7,6 +7,7 @@ import math
 
 import networkx as nx
 import numpy as np
+import plotly.colors as pc
 import plotly.graph_objects as go
 
 from tdamapper.core import ATTR_SIZE, aggregate_graph
@@ -53,6 +54,7 @@ def plot_plotly(
 ):
     node_col = aggregate_graph(colors, mapper_plot.graph, agg)
     fig = _figure(mapper_plot, node_col, node_size, width, height, title, cmap)
+    _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
     return fig
 
 
@@ -80,6 +82,7 @@ def plot_plotly_update(
         _update_traces_title(mapper_plot, fig, title)
     if (width is not None) and (height is not None):
         _update_layout(fig, width, height)
+    _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
     return fig
 
 
@@ -284,12 +287,11 @@ def _layout(width, height):
         title="",
     )
     return go.Layout(
-        uirevision="constant",
         plot_bgcolor="rgba(0, 0, 0, 0)",
         autosize=False,
         showlegend=False,
         hovermode="closest",
-        margin=dict(b=10, l=10, r=10, t=10),
+        margin=dict(b=5, l=5, r=5, t=5),
         xaxis=axis,
         yaxis=axis,
         width=width,
@@ -299,6 +301,197 @@ def _layout(width, height):
             yaxis=scene_axis,
             zaxis=scene_axis,
         ),
+    )
+
+
+def _sanitize_cmap(cmap):
+    # TODO: convert python cmap string to plotlyjs supported name
+    return cmap
+
+
+def _add_ui_to_layout(mapper_plot, mapper_fig, colors, node_size, agg, cmap):
+    cmaps = [
+        _sanitize_cmap(cmap),
+        "Jet",
+        "Viridis",
+        "Cividis",
+        "Spectral",
+        "Portland",
+        "HSV",
+        "Twilight",
+    ]
+    if mapper_plot.dim == 2:
+        menu_cmap = _cmap_buttons_2d(cmaps)
+        menu_color = _color_button_2d(mapper_plot, mapper_fig, colors, agg)
+    elif mapper_plot.dim == 3:
+        menu_cmap = _cmap_buttons_3d(cmaps)
+        menu_color = _color_button_3d(mapper_plot, mapper_fig, colors, agg)
+    slider_size = _node_size_slider(mapper_plot, node_size)
+    mapper_fig.update_layout(
+        updatemenus=[menu_cmap, menu_color],
+        sliders=[slider_size],
+    )
+
+
+def _cmap_buttons_2d(cmaps):
+    return dict(
+        buttons=[
+            dict(
+                label=cmap,
+                method="restyle",
+                args=[
+                    {
+                        "marker.colorscale": [cmap],
+                        "marker.line.colorscale": [cmap],
+                    },
+                    [1],
+                ],
+            )
+            for cmap in cmaps
+        ],
+        x=0.2,
+        xanchor="left",
+        y=1.0,
+        yanchor="top",
+        direction="down",
+    )
+
+
+def _cmap_buttons_3d(cmaps):
+    return dict(
+        buttons=[
+            dict(
+                label=cmap,
+                method="restyle",
+                args=[
+                    {
+                        "marker.colorscale": [None, cmap],
+                        "marker.line.colorscale": [None, cmap],
+                        "line.colorscale": [cmap, None],
+                    },
+                    [0, 1],
+                ],
+            )
+            for cmap in cmaps
+        ],
+        x=0.2,
+        xanchor="left",
+        y=1.0,
+        yanchor="top",
+        direction="down",
+    )
+
+
+def _node_size_slider(mapper_plot, node_size):
+    steps = [
[…]
