### 2025-05-17 4efb86e
simplified buttons code
