+++ src/tdamapper/_plot_plotly.py
@@ -27,31 +27,16 @@ _EDGE_COLOR = "#777"
 _TICKS_NUM = 10
 
 
-def _node_pos_array(graph, dim, node_pos):
-    return tuple([node_pos[n][i] for n in graph.nodes()] for i in range(dim))
-
-
-def _edge_pos_array(graph, dim, node_pos):
-    edges_arr = tuple([] for i in range(dim))
-    for edge in graph.edges():
-        pos0, pos1 = node_pos[edge[0]], node_pos[edge[1]]
-        for i in range(dim):
-            edges_arr[i].append(pos0[i])
-            edges_arr[i].append(pos1[i])
-            edges_arr[i].append(None)
-    return edges_arr
-
-
 def plot_plotly(
     mapper_plot,
-    width,
-    height,
-    title,
+    width: int,
+    height: int,
+    title: str,
     colors,
-    node_size=1,
+    node_size: int = 1,
     agg=np.nanmean,
-    cmap="jet",
-):
+    cmap: str = "Jet",
+) -> go.Figure:
     node_col = aggregate_graph(colors, mapper_plot.graph, agg)
     fig = _figure(mapper_plot, node_col, node_size, width, height, title, cmap)
     _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
@@ -60,15 +45,15 @@ def plot_plotly(
 
 def plot_plotly_update(
     mapper_plot,
-    fig,
-    width=None,
-    height=None,
-    title=None,
+    fig: go.Figure,
+    width: int | None = None,
+    height: int | None = None,
+    title: str | None = None,
     colors=None,
-    node_size=None,
+    node_size: int | None = None,
     agg=None,
-    cmap=None,
-):
+    cmap: str | None = None,
+) -> go.Figure:
     if (
         (colors is not None)
         and (node_size is not None)
@@ -84,6 +69,21 @@ def plot_plotly_update(
         _update_layout(fig, width, height)
     _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
     return fig
+
+
+def _node_pos_array(graph, dim, node_pos):
+    return tuple([node_pos[n][i] for n in graph.nodes()] for i in range(dim))
+
+
+def _edge_pos_array(graph, dim, node_pos):
+    edges_arr = tuple([] for i in range(dim))
+    for edge in graph.edges():
+        pos0, pos1 = node_pos[edge[0]], node_pos[edge[1]]
+        for i in range(dim):
+            edges_arr[i].append(pos0[i])
+            edges_arr[i].append(pos1[i])
+            edges_arr[i].append(None)
+    return edges_arr
 
 
 def _update_traces_col(mapper_plot, fig, colors, node_size, agg, cmap):
@@ -270,6 +270,7 @@ def _layout(width, height):
         showticklabels=False,
         showgrid=False,
         zeroline=False,
+        automargin=False,
         title="",
     )
     scene_axis = dict(
@@ -291,7 +292,7 @@ def _layout(width, height):
         autosize=False,
         showlegend=False,
         hovermode="closest",
-        margin=dict(b=5, l=5, r=5, t=5),
+        margin=dict(b=10, l=10, r=10, t=10),
         xaxis=axis,
         yaxis=axis,
         width=width,
@@ -320,12 +321,8 @@ def _add_ui_to_layout(mapper_plot, mapper_fig, colors, node_size, agg, cmap):
         "HSV",
         "Twilight",
     ]
-    if mapper_plot.dim == 2:
-        menu_cmap = _cmap_buttons_2d(cmaps)
[…]
