### 2025-05-18 26a4b94
Added first implementation of type hints
