### 2025-05-18 5ba6732
Refectored plot construction
