+++ src/tdamapper/_plot_plotly.py
@@ -4,6 +4,7 @@ plotly.
 """
 
 import math
+from typing import Optional
 
 import networkx as nx
 import numpy as np
@@ -26,19 +27,25 @@ _EDGE_COLOR = "#777"
 
 _TICKS_NUM = 10
 
+NODE_SIZE = 1
+
+CMAP = "Jet"
+
+TITLE = ""
+
 
 def plot_plotly(
     mapper_plot,
     width: int,
     height: int,
-    title: str,
-    colors,
-    node_size: int = 1,
+    title: str = TITLE,
+    node_size: int = NODE_SIZE,
+    colors=None,
     agg=np.nanmean,
-    cmap: str = "Jet",
+    cmap: str = CMAP,
 ) -> go.Figure:
-    node_col = aggregate_graph(colors, mapper_plot.graph, agg)
-    fig = _figure(mapper_plot, node_col, node_size, width, height, title, cmap)
+    # node_col = aggregate_graph(colors, mapper_plot.graph, agg)
+    fig = _figure(mapper_plot, width, height, title, node_size, colors, agg, cmap)
     _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
     return fig
 
@@ -46,28 +53,26 @@ def plot_plotly(
 def plot_plotly_update(
     mapper_plot,
     fig: go.Figure,
-    width: int | None = None,
-    height: int | None = None,
-    title: str | None = None,
+    width: Optional[int] = None,
+    height: Optional[int] = None,
+    title: Optional[str] = None,
+    node_size: Optional[int] = None,
     colors=None,
-    node_size: int | None = None,
     agg=None,
-    cmap: str | None = None,
+    cmap: Optional[str] = None,
 ) -> go.Figure:
-    if (
-        (colors is not None)
-        and (node_size is not None)
-        and (agg is not None)
-        and (cmap is not None)
-    ):
-        _update_traces_col(mapper_plot, fig, colors, node_size, agg, cmap)
-    if cmap is not None:
-        _update_traces_cmap(mapper_plot, fig, cmap)
-    if title is not None:
-        _update_traces_title(mapper_plot, fig, title)
     if (width is not None) and (height is not None):
         _update_layout(fig, width, height)
-    _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
+    if title is not None:
+        _set_title(mapper_plot, fig, title)
+    if node_size is not None:
+        _set_node_size(mapper_plot, fig, node_size)
+    if (colors is not None) and (agg is not None):
+        _set_colors(mapper_plot, fig, colors, agg)
+    if cmap is not None:
+        _set_cmap(mapper_plot, fig, cmap)
+    # _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
+    # TODO: understand how to update this
     return fig
 
 
@@ -86,35 +91,6 @@ def _edge_pos_array(graph, dim, node_pos):
     return edges_arr
 
 
-def _update_traces_col(mapper_plot, fig, colors, node_size, agg, cmap):
-    if (colors is not None) and (node_size is not None) and (agg is not None):
-        nodes_col = aggregate_graph(colors, mapper_plot.graph, agg)
-        colors_list = list(nodes_col.values())
-        _update_node_trace_col(mapper_plot, fig, nodes_col, colors_list, node_size)
-        _update_edge_trace_col(mapper_plot, fig, cmap, nodes_col, colors_list)
-
-
-def _update_edge_trace_col(mapper_plot, fig, cmap, colors_agg, colors_list):
-    colors_avg = []
-    for edge in mapper_plot.graph.edges():
-        c0, c1 = colors_agg[edge[0]], colors_agg[edge[1]]
-        colors_avg.append(c0)
[…]
