### 2025-05-18 8a3b335
Fixed colorscales using rgb conversion
