+++ src/tdamapper/_plot_plotly.py
@@ -4,7 +4,7 @@ plotly.
 """
 
 import math
-from typing import Optional
+from typing import Optional, Union
 
 import networkx as nx
 import numpy as np
@@ -38,6 +38,21 @@ DEFAULT_CMAP = "Jet"
 DEFAULT_TITLE = ""
 
 
+def _get_plotly_colorscales():
+    modules = ["sequential", "diverging", "cyclical"]
+    all_scales = []
+    for mod in modules:
+        m = getattr(pc, mod)
+        # keep only public scale names (start with uppercase letter)
+        valid = [name for name in dir(m) if name[0].isupper()]
+        all_scales.extend(valid)
+    cmaps = sorted(set(all_scales))
+    return {c.lower(): c for c in cmaps}
+
+
+PLOTLY_CMAPS = _get_plotly_colorscales()
+
+
 def plot_plotly(
     mapper_plot,
     width: int,
@@ -46,10 +61,14 @@ def plot_plotly(
     node_size: int = DEFAULT_NODE_SIZE,
     colors=None,
     agg=np.nanmean,
-    cmap: str = DEFAULT_CMAP,
+    cmap: Union[str, list[str]] = DEFAULT_CMAP,
 ) -> go.Figure:
-    fig = _figure(mapper_plot, width, height, title, node_size, colors, agg, cmap)
-    _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmap)
+    cmaps = [cmap] if isinstance(cmap, str) else cmap
+    colors = np.array(colors)
+    if colors.ndim == 1:
+        colors = colors.reshape(-1, 1)
+    fig = _figure(mapper_plot, width, height, title, node_size, colors, agg, cmaps)
+    _add_ui_to_layout(mapper_plot, fig, colors, node_size, agg, cmaps)
     return fig
 
 
@@ -160,7 +179,6 @@ def _set_colors(mapper_plot, fig, colors, agg):
             colors_avg.append(c1)
         fig.update_traces(
             patch=dict(
-                text=scatter_text,
                 marker=dict(
                     line_color=colors_avg,
                     line_cmin=min(colors_arr, default=None),
@@ -196,7 +214,7 @@ def _update_layout(fig, width, height):
     )
 
 
-def _figure(mapper_plot, width, height, title, node_size, colors, agg, cmap):
+def _figure(mapper_plot, width, height, title, node_size, colors, agg, cmaps):
     node_pos = mapper_plot.positions
     node_pos_arr = _node_pos_array(
         mapper_plot.graph,
@@ -213,8 +231,8 @@ def _figure(mapper_plot, width, height, title, node_size, colors, agg, cmap):
     _layout_ = _layout(width, height)
     fig = go.Figure(data=[_edges_tr, _nodes_tr], layout=_layout_)
 
-    _set_cmap(mapper_plot, fig, cmap)
-    _set_colors(mapper_plot, fig, colors, agg)
+    _set_cmap(mapper_plot, fig, cmaps[0])
+    _set_colors(mapper_plot, fig, colors[:, 0], agg)
     _set_node_size(mapper_plot, fig, node_size)
     _set_title(mapper_plot, fig, title)
 
@@ -363,45 +381,15 @@ def _layout(width, height):
     )
 
 
-def _get_plotly_js_colorscales():
-    """
-    Return a sorted list of all case-sensitive color-scale names
-    supported by Plotly.js (via the Python plotly.colors wrapper).
-    Includes sequential, diverging, and cyclical scales.
-    """
-    modules = ["sequential", "diverging", "cyclical"]
-    all_scales = []
-    for mod in modules:
-        m = getattr(pc, mod)
-        # keep only public scale names (start with uppercase letter)
-        valid = [name for name in dir(m) if name[0].isupper()]
[…]
