### 2025-05-19 f851822
Fixed type hints for python 3.8
