+++ src/tdamapper/plot.py
@@ -153,6 +153,7 @@ class MapperPlot:
     def plot_plotly(
         self,
         colors,
+        color_names,
         node_size=1,
         agg=np.nanmean,
         title=None,
@@ -167,6 +168,9 @@ class MapperPlot:
             node in the graph, useful for highlighting different features of
             the data.
         :type colors: array-like of shape (n,) or list-like of size n
+        :param color_names: A list of color names that will be used in the UI
+            buttons to represent and chose the rendered color.
+        :type color_names: list[str] of size n
         :param node_size: A scaling factor for node size. Defaults to 1.
         :type node_size: float, optional
         :param agg: A function used to aggregate the `colors` array over the
@@ -194,6 +198,7 @@ class MapperPlot:
         return plot_plotly(
             self,
             colors=colors,
+            color_names=color_names,
             node_size=node_size,
             agg=agg,
             title=title,
