+++ app/streamlit_app.py
@@ -177,14 +177,6 @@ def _fix_data(data):
     return df
 
 
-def _get_dim(fig):
-    dim = 2
-    for trace in fig.data:
-        if "3d" in trace.type:
-            dim = 3
-    return dim
-
-
 def _get_graph_no_attribs(graph):
     graph_no_attribs = nx.Graph()
     graph_no_attribs.add_nodes_from(graph.nodes())
@@ -573,36 +565,6 @@ def plot_agg_input_section():
     return agg, agg_name
 
 
-def plot_cmap_input_section():
-    cmap_type = st.selectbox(
-        "Colormap",
-        options=[
-            V_CMAP_JET,
-            V_CMAP_VIRIDIS,
-            V_CMAP_CIVIDIS,
-            V_CMAP_PORTLAND,
-            V_CMAP_SPECTRAL,
-            V_CMAP_HSV,
-            V_CMAP_TWILIGHT,
-        ],
-    )
-    cmap = V_CMAPS.get(cmap_type, "Jet")
-    return cmap
-
-
-def plot_color_input_section(df_X, df_y):
-    X_cols = list(df_X.columns)
-    y_cols = list(df_y.columns)
-    col_feat = st.selectbox(
-        "Color",
-        options=y_cols + X_cols,
-    )
-    if col_feat in X_cols:
-        return df_X[col_feat].to_numpy(), col_feat
-    elif col_feat in y_cols:
-        return df_y[col_feat].to_numpy(), col_feat
-
-
 @st.cache_data(
     hash_funcs={
         "networkx.classes.graph.Graph": lambda g: _encode_graph(
@@ -652,41 +614,35 @@ def mapper_plot_section(mapper_graph):
     hash_funcs={"tdamapper.plot.MapperPlot": lambda mp: mp.positions},
     show_spinner="Rendering Mapper",
 )
-def compute_mapper_fig(
-    mapper_plot, colors, node_size, cmap, _agg, agg_name, colors_feat
-):
+def compute_mapper_fig(mapper_plot, colors, node_size, cmap, _agg, agg_name):
     logger.info("Generating Mapper figure")
     mapper_fig = mapper_plot.plot_plotly(
         colors,
+        color_names=colors.columns,
         node_size=node_size,
         agg=_agg,
-        title=f"{agg_name} of {colors_feat}",
+        title=agg_name,
         cmap=cmap,
         width=600,
         height=600,
     )
-    mapper_fig.update_layout(uirevision="constant")
     return mapper_fig
 
 
 def mapper_figure_section(df_X, df_y, mapper_plot):
     st.header("🎨 Plot")
     agg, agg_name = plot_agg_input_section()
-    colors, colors_feat = plot_color_input_section(df_X, df_y)
+    colors = pd.concat([df_y, df_X], axis=1)
     mapper_fig = compute_mapper_fig(
         mapper_plot,
         colors=colors,
         node_size=1.0,
         _agg=agg,
-        cmap="Jet",
+        cmap=["Jet", "Viridis", "Cividis"],
         agg_name=agg_name,
-        colors_feat=colors_feat,
     )
-    dim = _get_dim(mapper_fig)
     mapper_fig.update_layout(
-        dragmode="orbit" if dim == 3 else "pan",
-        uirevision="constant",
-        margin=dict(b=0, l=0, r=0, t=0),
+        margin=dict(b=5, l=5, r=5, t=5),
     )
     mapper_fig.update_xaxes(
         showline=False,
@@ -697,160 +653,18 @@ def mapper_figure_section(df_X, df_y, mapper_plot):
         scaleratio=1,
     )
 
[…]
