### 2025-05-20 42f1165
Fixed rendering of notebooks. Extended title argument for passing lists
