+++ src/tdamapper/plot.py
@@ -153,7 +153,6 @@ class MapperPlot:
     def plot_plotly(
         self,
         colors,
-        color_names=None,
         node_size=1,
         agg=np.nanmean,
         title=None,
@@ -168,9 +167,6 @@ class MapperPlot:
             node in the graph, useful for highlighting different features of
             the data.
         :type colors: array-like of shape (n,) or list-like of size n
-        :param color_names: A list of color names that will be used in the UI
-            buttons to represent and chose the rendered color.
-        :type color_names: list[str] of size n, optional
         :param node_size: A scaling factor for node size. Defaults to 1.
         :type node_size: float, optional
         :param agg: A function used to aggregate the `colors` array over the
@@ -178,8 +174,10 @@ class MapperPlot:
             obtained by mapping the aggregated value with the colormap `cmap`.
             Defaults to `numpy.nanmean`.
         :type agg: Callable, optional
-        :param title: The title to be displayed alongside the figure.
-        :type title: str, optional
+        :param title: The title for the colormap. When colors has shape (n, m)
+            and title is a list of string, each item will be used as title for
+            its corresponding colormap.
+        :type title: str, list[str], optional
         :param width: The desired width of the figure in pixels. Defaults to
             512.
         :type width: int, optional
@@ -198,7 +196,6 @@ class MapperPlot:
         return plot_plotly(
             self,
             colors=colors,
-            color_names=color_names,
             node_size=node_size,
             agg=agg,
             title=title,
