### 2025-05-21 3e07f19
Added profiling decorator. Improved tests
