### 2025-05-21 fb05faf
Refactored flat vptree using numba for array operations
