+++ src/tdamapper/utils/vptree_flat.py
@@ -1,9 +1,13 @@
 from random import randrange
 
+import numpy as np
+from numba import njit
+
 from tdamapper.utils.heap import MaxHeap
 from tdamapper.utils.metrics import get_metric
 
 
+@njit
 def _swap(arr, i, j):
     arr[i], arr[j] = arr[j], arr[i]
 
@@ -12,25 +16,39 @@ def _mid(start, end):
     return (start + end) // 2
 
 
-def _partition(data, start, end, p_ord):
+@njit
+def _partition(distances, indices, is_terminal, start, end, p_ord):
     higher = start
     for j in range(start, end):
-        j_ord, _, _ = data[j]
+        j_ord = distances[j]
         if j_ord < p_ord:
-            _swap(data, higher, j)
+            _swap(distances, higher, j)
+            _swap(indices, higher, j)
+            _swap(is_terminal, higher, j)
             higher += 1
     return higher
 
 
-def _quickselect(data, start, end, k):
+def _quickselect(distances, indices, is_terminal, start, end, k):
     if (k < start) or (k >= end):
         return
     start_, end_, higher = start, end, None
     while higher != k + 1:
-        p, _, _ = data[k]
-        _swap(data, start_, k)
-        higher = _partition(data, start_ + 1, end_, p)
-        _swap(data, start_, higher - 1)
+        # TODO: pivot_index = randrange(start_, end_)
+        pivot_index = k
+
+        p = distances[pivot_index]
+
+        _swap(distances, start_, pivot_index)
+        _swap(indices, start_, pivot_index)
+        _swap(is_terminal, start_, pivot_index)
+
+        higher = _partition(distances, indices, is_terminal, start_ + 1, end_, p)
+
+        _swap(distances, start_, higher - 1)
+        _swap(indices, start_, higher - 1)
+        _swap(is_terminal, start_, higher - 1)
+
         if k <= higher - 1:
             end_ = higher
         else:
@@ -53,7 +71,9 @@ class VPTree:
         self.__leaf_capacity = leaf_capacity
         self.__leaf_radius = leaf_radius
         self.__pivoting = pivoting
-        self.__dataset = self._Build(self, X).build()
+        self.__dataset, self.__distances, self.__indices, self.__is_terminal = (
+            self._Build(self, X).build()
+        )
 
     def get_metric(self):
         return self.__metric
@@ -77,11 +97,25 @@ class VPTree:
         metric_params = self.__metric_params or {}
         return get_metric(self.__metric, **metric_params)
 
+    def _get_distances(self):
+        return self.__distances
+
+    def _get_indices(self):
+        return self.__indices
+
+    def _get_is_terminal(self):
+        return self.__is_terminal
+
     class _Build:
 
         def __init__(self, vpt, X):
             self.__distance = vpt._get_distance()
-            self.__dataset = [(0.0, x, False) for x in X]
+
+            self.__dataset = [x for x in X]
+            self.__indices = np.array([i for i in range(len(self.__dataset))])
+            self.__distances = np.array([0.0 for _ in X])
+            self.__is_terminal = np.array([False for _ in X])
+
             self.__leaf_capacity = vpt.get_leaf_capacity()
             self.__leaf_radius = vpt.get_leaf_radius()
             pivoting = vpt.get_pivoting()
@@ -99,14 +133,21 @@ class VPTree:
                 return
             pivot = randrange(start, end)
             if pivot > start:
-                _swap(self.__dataset, start, pivot)
+                _swap(self.__distances, start, pivot)
+                _swap(self.__indices, start, pivot)
+                _swap(self.__is_terminal, start, pivot)
 
         def _furthest(self, start, end, i):
             furthest_dist = 0.0
             furthest = start
[…]
