+++ src/tdamapper/utils/vptree_hier.py
@@ -1,12 +1,10 @@
 from random import randrange
 
+import numpy as np
+
 from tdamapper.utils.heap import MaxHeap
 from tdamapper.utils.metrics import get_metric
-from tdamapper.utils.quickselect import quickselect
-
-
-def _swap(arr, i, j):
-    arr[i], arr[j] = arr[j], arr[i]
+from tdamapper.utils.quickselect import quickselect, swap_all
 
 
 def _mid(start, end):
@@ -29,8 +27,13 @@ class VPTree:
         self.__leaf_capacity = leaf_capacity
         self.__leaf_radius = leaf_radius
         self.__pivoting = pivoting
-        tree, dataset = self._Build(self, X).build()
-        self.__tree, self.__dataset = tree, dataset
+        tree, dataset, distances, indices = self._Build(self, X).build()
+        self.__tree, self.__dataset, self.__distances, self.__indices = (
+            tree,
+            dataset,
+            distances,
+            indices,
+        )
 
     def get_metric(self):
         return self.__metric
@@ -53,6 +56,12 @@ class VPTree:
     def _get_dataset(self):
         return self.__dataset
 
+    def _get_distances(self):
+        return self.__distances
+
+    def _get_indices(self):
+        return self.__indices
+
     def _get_distance(self):
         metric_params = self.__metric_params or {}
         return get_metric(self.__metric, **metric_params)
@@ -61,7 +70,10 @@ class VPTree:
 
         def __init__(self, vpt, X):
             self.__distance = vpt._get_distance()
-            self.__dataset = [(0.0, x) for x in X]
+            self.__dataset = [x for x in X]
+            self.__indices = np.array([i for i in range(len(self.__dataset))])
+            self.__distances = np.array([0.0 for _ in X])
+
             self.__leaf_capacity = vpt.get_leaf_capacity()
             self.__leaf_radius = vpt.get_leaf_radius()
             pivoting = vpt.get_pivoting()
@@ -79,14 +91,20 @@ class VPTree:
                 return
             pivot = randrange(start, end)
             if pivot > start:
-                _swap(self.__dataset, start, pivot)
+                swap_all(self.__distances, start, pivot, self.__indices)
+
+        def _get_point(self, i):
+            return self.__dataset[self.__indices[i]]
+
+        def _get_dist(self, i):
+            return self.__distances[i]
 
         def _furthest(self, start, end, i):
             furthest_dist = 0.0
             furthest = start
-            _, i_point = self.__dataset[i]
+            i_point = self._get_point(i)
             for j in range(start, end):
-                _, j_point = self.__dataset[j]
+                j_point = self._get_point(j)
                 j_dist = self.__distance(i_point, j_point)
                 if j_dist > furthest_dist:
                     furthest = j
@@ -100,26 +118,26 @@ class VPTree:
             furthest_rnd = self._furthest(start, end, rnd)
             furthest = self._furthest(start, end, furthest_rnd)
             if furthest > start:
-                _swap(self.__dataset, start, furthest)
+                swap_all(self.__distances, start, furthest, self.__indices)
 
         def _update(self, start, end):
             self.__pivoting(start, end)
-            _, v_point = self.__dataset[start]
+            v_point = self._get_point(start)
             for i in range(start + 1, end):
-                _, point = self.__dataset[i]
-                self.__dataset[i] = self.__distance(v_point, point), point
+                point = self._get_point(i)
+                self.__distances[i] = self.__distance(v_point, point)
 
         def build(self):
             tree = self._build_rec(0, len(self.__dataset))
-            return tree, self.__dataset
+            return tree, self.__dataset, self.__distances, self.__indices
 
         def _build_rec(self, start, end):
             mid = _mid(start, end)
             self._update(start, end)
[…]
