### 2025-05-22 4b6087b
Refactor of vptree code. Improved modularity and performance using numba
