+++ src/tdamapper/utils/vptree_hier.py
@@ -1,290 +1,0 @@
-from random import randrange
-
-import numpy as np
-
-from tdamapper.utils.heap import MaxHeap
-from tdamapper.utils.metrics import get_metric
-from tdamapper.utils.quickselect import quickselect, swap_all
-
-
-def _mid(start, end):
-    return (start + end) // 2
-
-
-class VPTree:
-
-    def __init__(
-        self,
-        X,
-        metric="euclidean",
-        metric_params=None,
-        leaf_capacity=1,
-        leaf_radius=0.0,
-        pivoting=None,
-    ):
-        self.__metric = metric
-        self.__metric_params = metric_params
-        self.__leaf_capacity = leaf_capacity
-        self.__leaf_radius = leaf_radius
-        self.__pivoting = pivoting
-        tree, dataset, distances, indices = self._Build(self, X).build()
-        self.__tree, self.__dataset, self.__distances, self.__indices = (
-            tree,
-            dataset,
-            distances,
-            indices,
-        )
-
-    def get_metric(self):
-        return self.__metric
-
-    def get_metric_params(self):
-        return self.__metric_params
-
-    def get_leaf_capacity(self):
-        return self.__leaf_capacity
-
-    def get_leaf_radius(self):
-        return self.__leaf_radius
-
-    def get_pivoting(self):
-        return self.__pivoting
-
-    def _get_tree(self):
-        return self.__tree
-
-    def _get_dataset(self):
-        return self.__dataset
-
-    def _get_distances(self):
-        return self.__distances
-
-    def _get_indices(self):
-        return self.__indices
-
-    def _get_distance(self):
-        metric_params = self.__metric_params or {}
-        return get_metric(self.__metric, **metric_params)
-
-    class _Build:
-
-        def __init__(self, vpt, X):
-            self.__distance = vpt._get_distance()
-            self.__dataset = [x for x in X]
-            self.__indices = np.array([i for i in range(len(self.__dataset))])
-            self.__distances = np.array([0.0 for _ in X])
-
-            self.__leaf_capacity = vpt.get_leaf_capacity()
-            self.__leaf_radius = vpt.get_leaf_radius()
-            pivoting = vpt.get_pivoting()
-            self.__pivoting = self._pivoting_disabled
-            if pivoting == "random":
-                self.__pivoting = self._pivoting_random
-            elif pivoting == "furthest":
-                self.__pivoting = self._pivoting_furthest
-
-        def _pivoting_disabled(self, start, end):
-            pass
-
-        def _pivoting_random(self, start, end):
-            if end <= start:
-                return
-            pivot = randrange(start, end)
-            if pivot > start:
-                swap_all(self.__distances, start, pivot, self.__indices)
-
-        def _get_point(self, i):
-            return self.__dataset[self.__indices[i]]
-
-        def _get_dist(self, i):
-            return self.__distances[i]
-
-        def _furthest(self, start, end, i):
-            furthest_dist = 0.0
-            furthest = start
-            i_point = self._get_point(i)
-            for j in range(start, end):
-                j_point = self._get_point(j)
-                j_dist = self.__distance(i_point, j_point)
-                if j_dist > furthest_dist:
-                    furthest = j
-                    furthest_dist = j_dist
-            return furthest
-
-        def _pivoting_furthest(self, start, end):
-            if end <= start:
-                return
-            rnd = randrange(start, end)
-            furthest_rnd = self._furthest(start, end, rnd)
-            furthest = self._furthest(start, end, furthest_rnd)
-            if furthest > start:
-                swap_all(self.__distances, start, furthest, self.__indices)
-
-        def _update(self, start, end):
-            self.__pivoting(start, end)
-            v_point = self._get_point(start)
-            for i in range(start + 1, end):
-                point = self._get_point(i)
[…]
