### 2025-05-22 262b75b
Removed variadic args to support numba on python < 3.10
