### 2025-05-22 8c8313f
Added flags for compatibility of numba with python 3.8
