### 2025-05-22 fa915da
Removed None for compatibility with numba
