### 2025-05-22 e8603cd
Skipped njitted functions from coverage count
