+++ src/tdamapper/utils/_metrics.py
@@ -2,27 +2,27 @@ import numpy as np
 from numba import njit
 
 
-@njit(fastmath=True)
+@njit(fastmath=True)  # pragma: no cover
 def euclidean(x, y):
     return np.linalg.norm(x - y)
 
 
-@njit(fastmath=True)
+@njit(fastmath=True)  # pragma: no cover
 def manhattan(x, y):
     return np.linalg.norm(x - y, ord=1)
 
 
-@njit(fastmath=True)
+@njit(fastmath=True)  # pragma: no cover
 def chebyshev(x, y):
     return np.linalg.norm(x - y, ord=np.inf)
 
 
-@njit(fastmath=True)
+@njit(fastmath=True)  # pragma: no cover
 def minkowski(p, x, y):
     return np.linalg.norm(x - y, ord=p)
 
 
-@njit(fastmath=True)
+@njit(fastmath=True)  # pragma: no cover
 def cosine(x, y):
     xy = np.dot(x, y)
     xx = np.linalg.norm(x)
