### 2025-05-22 1bf67c9
Skipped njitted functions from coverage
