### 2025-05-24 e13162e
Refactored tests to use pytest instead of unittest
