+++ tests/test_unit_plot.py
@@ -1,5 +1,3 @@
-import unittest
-
 import networkx as nx
 import numpy as np
 
@@ -9,111 +7,111 @@ from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
 
 
-class TestMapperPlot(unittest.TestCase):
+def test_two_connected_clusters():
+    data = [
+        np.array([0.0, 1.0]),
+        np.array([1.0, 0.0]),
+        np.array([0.0, 0.0]),
+        np.array([1.0, 1.0]),
+    ]
+    mp = MapperAlgorithm(
+        cover=BallCover(1.1, metric="euclidean"), clustering=TrivialClustering()
+    )
+    g = mp.fit_transform(data, data)
+    mp_plot1 = MapperPlot(g, dim=2, seed=123, iterations=10)
+    mp_plot1.plot_plotly(
+        colors=data,
+        agg=np.nanmax,
+        width=200,
+        height=200,
+        title="example",
+        cmap="jet",
+    )
+    mp_plot2 = MapperPlot(g, dim=3, seed=123, iterations=10)
+    fig2 = mp_plot2.plot_plotly(
+        colors=data,
+        agg=np.nanmax,
+        width=200,
+        height=200,
+        title="example",
+        cmap="jet",
+    )
+    mp_plot2.plot_plotly_update(
+        fig2,
+        colors=data,
+        agg=np.nanmin,
+        width=300,
+        height=300,
+        title="example-updated",
+        cmap="viridis",
+    )
+    mp_plot3 = MapperPlot(g, dim=2)
+    mp_plot3.plot_matplotlib(width=300, height=300, colors=data)
+    mp_plot3.plot_pyvis(
+        width=512,
+        height=512,
+        colors=data,
+        output_file="network.html",
+    )
 
-    def test_two_connected_clusters(self):
-        data = [
-            np.array([0.0, 1.0]),
-            np.array([1.0, 0.0]),
-            np.array([0.0, 0.0]),
-            np.array([1.0, 1.0]),
-        ]
-        mp = MapperAlgorithm(
-            cover=BallCover(1.1, metric="euclidean"), clustering=TrivialClustering()
-        )
-        g = mp.fit_transform(data, data)
-        mp_plot1 = MapperPlot(g, dim=2, seed=123, iterations=10)
-        mp_plot1.plot_plotly(
-            colors=data,
-            agg=np.nanmax,
-            width=200,
-            height=200,
-            title="example",
-            cmap="jet",
-        )
-        mp_plot2 = MapperPlot(g, dim=3, seed=123, iterations=10)
-        fig2 = mp_plot2.plot_plotly(
-            colors=data,
-            agg=np.nanmax,
-            width=200,
-            height=200,
-            title="example",
-            cmap="jet",
-        )
-        mp_plot2.plot_plotly_update(
-            fig2,
-            colors=data,
-            agg=np.nanmin,
-            width=300,
-            height=300,
-            title="example-updated",
-            cmap="viridis",
-        )
-        mp_plot3 = MapperPlot(g, dim=2)
-        mp_plot3.plot_matplotlib(width=300, height=300, colors=data)
-        mp_plot3.plot_pyvis(
-            width=512,
-            height=512,
-            colors=data,
-            output_file="network.html",
-        )
 
-    def test_empty_graph(self):
-        empty_graph = nx.Graph()
-        mapper_plot = MapperPlot(empty_graph, dim=2)
-        mapper_plot.plot_matplotlib(colors=[])
-        mapper_plot.plot_plotly(colors=[])
-        mapper_plot.plot_pyvis(colors=[], output_file="tmp.html")
+def test_empty_graph():
+    empty_graph = nx.Graph()
[…]
