### 2025-05-26 7534138
Improved code. Minor fixes
