### 2025-05-28 859a38d
Added embedded rendering of plotly figure to prevent view reset
