### 2025-05-29 a5d2674
First implementation using nicegui
