+++ app/nicegui_app.py
@@ -1,0 +1,309 @@
+import numpy as np
+import plotly.graph_objs as go
+from nicegui import ui
+from nicegui.events import ValueChangeEventArguments
+from sklearn.cluster import AgglomerativeClustering, KMeans
+from sklearn.datasets import load_digits, make_circles
+from sklearn.decomposition import PCA
+from umap import UMAP
+
+from tdamapper.core import TrivialClustering, TrivialCover
+from tdamapper.cover import BallCover, CubicalCover, KNNCover
+from tdamapper.learn import MapperAlgorithm
+from tdamapper.plot import MapperPlot
+
+
+def mode(arr):
+    values, counts = np.unique(arr, return_counts=True)
+    max_count = np.max(counts)
+    mode_values = values[counts == max_count]
+    return np.nanmean(mode_values)
+
+
+def _identity(X):
+    return X
+
+
+def _pca(n_components):
+    pca = PCA(n_components=n_components, random_state=42)
+
+    def _func(X):
+        return pca.fit_transform(X)
+
+    return _func
+
+
+def _umap(n_components):
+    um = UMAP(n_components=n_components, random_state=42)
+
+    def _func(X):
+        return um.fit_transform(X)
+
+    return _func
+
+
+class App:
+
+    def build_lens(self):
+        self.opt_lens_id = "Identity"
+        self.opt_lens_pca = "PCA"
+        self.opt_lens_umap = "UMAP"
+
+        self.lens_type = ui.select(
+            label="Lens type",
+            options=[
+                self.opt_lens_id,
+                self.opt_lens_pca,
+                self.opt_lens_umap,
+            ],
+            value=self.opt_lens_pca,
+            on_change=self.update,
+        ).classes("w-full")
+        self.pca_n_components = ui.number(
+            label="PCA Components",
+            min=1,
+            max=10,
+            value=2,
+            on_change=self.update,
+        ).classes("w-full")
+        self.pca_n_components.bind_visibility_from(
+            target_object=self.lens_type,
+            target_name="value",
+            value=self.opt_lens_pca,
+        )
+        self.umap_n_components = ui.number(
+            label="UMAP Components",
+            min=1,
+            max=10,
+            value=2,
+            on_change=self.update,
+        ).classes("w-full")
+        self.umap_n_components.bind_visibility_from(
+            target_object=self.lens_type,
+            target_name="value",
+            value=self.opt_lens_umap,
+        )
+
+    def build_cover(self):
+        self.opt_cover_trivial = "Trivial"
+        self.opt_cover_cubical = "Cubical"
+        self.opt_cover_ball = "Ball"
+        self.opt_cover_knn = "KNN"
+
+        self.cover_type = ui.select(
+            label="Cover type",
+            options=[
+                self.opt_cover_trivial,
+                self.opt_cover_cubical,
+                self.opt_cover_ball,
+                self.opt_cover_knn,
+            ],
+            value=self.opt_cover_cubical,
+            on_change=self.update,
+        ).classes("w-full")
+        self.cover_cubical_n = ui.number(
+            label="Intervals",
+            min=1,
+            max=10,
+            value=2,
+            on_change=self.update,
+        ).classes("w-full")
+        self.cover_cubical_n.bind_visibility_from(
+            target_object=self.cover_type,
+            target_name="value",
+            value=self.opt_cover_cubical,
+        )
+        self.cover_cubical_overlap = ui.number(
+            label="Overlap",
+            min=0.0,
+            max=1.0,
+            value=0.5,
+            on_change=self.update,
+        ).classes("w-full")
+        self.cover_cubical_overlap.bind_visibility_from(
+            target_object=self.cover_type,
+            target_name="value",
+            value=self.opt_cover_cubical,
+        )
[…]
