### 2025-05-30 ee5e8b5
Added ui elements for cover and clustering. Added async worker function for plot update
