+++ app/nicegui_app.py
@@ -1,9 +1,8 @@
 import numpy as np
 import plotly.graph_objs as go
-from nicegui import ui
-from nicegui.events import ValueChangeEventArguments
-from sklearn.cluster import AgglomerativeClustering, KMeans
-from sklearn.datasets import load_digits, make_circles
+from nicegui import run, ui
+from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
+from sklearn.datasets import load_digits
 from sklearn.decomposition import PCA
 from umap import UMAP
 
@@ -42,229 +41,238 @@ def _umap(n_components):
     return _func
 
 
+LENS_IDENTITY = "Identity"
+LENS_PCA = "PCA"
+LENS_UMAP = "UMAP"
+
+COVER_TRIVIAL = "Trivial"
+COVER_CUBICAL = "Cubical"
+COVER_BALL = "Ball"
+COVER_KNN = "KNN"
+
+CLUSTERING_TRIVIAL = "Trivial"
+CLUSTERING_KMEANS = "KMeans"
+CLUSTERING_AGGLOMERATIVE = "Agglomerative"
+CLUSTERING_DBSCAN = "DBSCAN"
+
+
 class App:
 
     def build_lens(self):
-        self.opt_lens_id = "Identity"
-        self.opt_lens_pca = "PCA"
-        self.opt_lens_umap = "UMAP"
-
         self.lens_type = ui.select(
             label="Lens type",
             options=[
-                self.opt_lens_id,
-                self.opt_lens_pca,
-                self.opt_lens_umap,
+                LENS_IDENTITY,
+                LENS_PCA,
+                LENS_UMAP,
             ],
-            value=self.opt_lens_pca,
-            on_change=self.update,
+            value=LENS_PCA,
+            on_change=self.update_handler,
         ).classes("w-full")
         self.pca_n_components = ui.number(
             label="PCA Components",
             min=1,
             max=10,
             value=2,
-            on_change=self.update,
+            on_change=self.update_handler,
         ).classes("w-full")
         self.pca_n_components.bind_visibility_from(
             target_object=self.lens_type,
             target_name="value",
-            value=self.opt_lens_pca,
+            value=LENS_PCA,
         )
         self.umap_n_components = ui.number(
             label="UMAP Components",
             min=1,
             max=10,
             value=2,
-            on_change=self.update,
+            on_change=self.update_handler,
         ).classes("w-full")
         self.umap_n_components.bind_visibility_from(
             target_object=self.lens_type,
             target_name="value",
-            value=self.opt_lens_umap,
+            value=LENS_UMAP,
         )
 
     def build_cover(self):
-        self.opt_cover_trivial = "Trivial"
-        self.opt_cover_cubical = "Cubical"
-        self.opt_cover_ball = "Ball"
-        self.opt_cover_knn = "KNN"
 
         self.cover_type = ui.select(
             label="Cover type",
             options=[
-                self.opt_cover_trivial,
-                self.opt_cover_cubical,
-                self.opt_cover_ball,
-                self.opt_cover_knn,
+                COVER_TRIVIAL,
+                COVER_CUBICAL,
+                COVER_BALL,
+                COVER_KNN,
             ],
-            value=self.opt_cover_cubical,
-            on_change=self.update,
+            value=COVER_CUBICAL,
+            on_change=self.update_handler,
         ).classes("w-full")
-        self.cover_cubical_n = ui.number(
+        self.cover_cubical_n_intervals = ui.number(
             label="Intervals",
             min=1,
-            max=10,
+            max=100,
             value=2,
-            on_change=self.update,
+            on_change=self.update_handler,
         ).classes("w-full")
[…]
