+++ app/nicegui_app.py
@@ -2,7 +2,7 @@ import numpy as np
 import plotly.graph_objs as go
 from nicegui import run, ui
 from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
-from sklearn.datasets import load_digits
+from sklearn.datasets import load_digits, load_iris
 from sklearn.decomposition import PCA
 from umap import UMAP
 
@@ -55,8 +55,62 @@ CLUSTERING_KMEANS = "KMeans"
 CLUSTERING_AGGLOMERATIVE = "Agglomerative"
 CLUSTERING_DBSCAN = "DBSCAN"
 
+DATA_SOURCE_EXAMPLE = "Example"
+DATA_SOURCE_CSV = "CSV"
+DATA_SOURCE_OPENML = "OpenML"
+
+DATA_SOURCE_EXAMPLE_DIGITS = "Digits"
+DATA_SOURCE_EXAMPLE_IRIS = "Iris"
+
+DRAW_3D = "3D"
+DRAW_2D = "2D"
+DRAW_ITERATIONS = 50
+
 
 class App:
+
+    def build_dataset(self):
+        self.data_source_type = ui.select(
+            label="Data Source",
+            options=[
+                DATA_SOURCE_EXAMPLE,
+                DATA_SOURCE_CSV,
+                DATA_SOURCE_OPENML,
+            ],
+            value=DATA_SOURCE_EXAMPLE,
+            on_change=self.update_dataset_handler,
+        ).classes("w-full")
+        self.data_source_example_file = ui.select(
+            label="File",
+            options=[
+                DATA_SOURCE_EXAMPLE_DIGITS,
+                DATA_SOURCE_EXAMPLE_IRIS,
+            ],
+            value=DATA_SOURCE_EXAMPLE_DIGITS,
+            on_change=self.update_dataset_handler,
+        ).classes("w-full")
+        self.data_source_example_file.bind_visibility_from(
+            target_object=self.data_source_type,
+            target_name="value",
+            value=DATA_SOURCE_EXAMPLE,
+        )
+        self.data_source_csv = ui.upload(
+            on_upload=self.update_dataset_handler,
+        ).classes("w-full")
+        self.data_source_csv.bind_visibility_from(
+            target_object=self.data_source_type,
+            target_name="value",
+            value=DATA_SOURCE_CSV,
+        )
+        self.data_source_openml = ui.input(
+            label="OpenML Code",
+            on_change=self.update_dataset_handler,
+        ).classes("w-full")
+        self.data_source_openml.bind_visibility_from(
+            target_object=self.data_source_type,
+            target_name="value",
+            value=DATA_SOURCE_OPENML,
+        )
 
     def build_lens(self):
         self.lens_type = ui.select(
@@ -67,14 +121,14 @@ class App:
                 LENS_UMAP,
             ],
             value=LENS_PCA,
-            on_change=self.update_handler,
+            on_change=self.update_graph_handler,
         ).classes("w-full")
         self.pca_n_components = ui.number(
             label="PCA Components",
             min=1,
             max=10,
             value=2,
-            on_change=self.update_handler,
+            on_change=self.update_graph_handler,
         ).classes("w-full")
         self.pca_n_components.bind_visibility_from(
             target_object=self.lens_type,
@@ -86,7 +140,7 @@ class App:
             min=1,
             max=10,
             value=2,
-            on_change=self.update_handler,
+            on_change=self.update_graph_handler,
         ).classes("w-full")
         self.umap_n_components.bind_visibility_from(
             target_object=self.lens_type,
@@ -95,7 +149,6 @@ class App:
         )
 
     def build_cover(self):
-
         self.cover_type = ui.select(
             label="Cover type",
[…]
