### 2025-06-01 fae1c42
Improved attribute names. Improved ui handling
