### 2025-06-01 6887d24
Added type hints. Improved handling of missing arguments
