+++ src/tdamapper/plot_backends/plot_plotly.py
@@ -60,7 +60,9 @@ class PlotlyUI:
         self.menu_color = None
         self.slider_size = None
 
-    def set_menu_cmap(self, mapper_plot, cmaps):
+    def set_menu_cmap(self, mapper_plot, cmaps: Optional[List[str]]) -> None:
+        if cmaps is None:
+            return
         cmaps_plotly = [PLOTLY_CMAPS.get(c.lower()) for c in cmaps]
         self.menu_cmap = _ui_cmap(mapper_plot, cmaps_plotly)
 
@@ -71,8 +73,10 @@ class PlotlyUI:
         self.slider_size = _ui_node_size(mapper_plot, node_sizes)
 
 
-def _to_cmaps(cmap: Union[str, List[str]]) -> List[str]:
+def _to_cmaps(cmap: Optional[Union[str, List[str]]]) -> List[str]:
     """Convert a single cmap or a list of cmaps to a list of cmaps."""
+    if cmap is None:
+        return [DEFAULT_CMAP]
     if isinstance(cmap, str):
         return [cmap]
     elif isinstance(cmap, list):
@@ -94,11 +98,11 @@ def _to_colors(colors: Union[np.ndarray, List[float]]) -> np.ndarray:
         )
 
 
-def _to_titles(title, colors_num):
+def _to_titles(title: Optional[Union[str, List[str]]], colors_num: int) -> List[str]:
     if title is None:
-        return [DEFAULT_TITLE for _ in range(colors_num)]
+        return [f"{i}" for i in range(colors_num)]
     elif isinstance(title, str):
-        return [title for _ in range(colors_num)]
+        return [f"{title} {i}" for i in range(colors_num)]
     elif isinstance(title, list) and len(title) == colors_num:
         return title
     else:
@@ -107,7 +111,9 @@ def _to_titles(title, colors_num):
         )
 
 
-def _to_node_sizes(node_size):
+def _to_node_sizes(
+    node_size: Optional[Union[int, float, List[Union[int, float]]]]
+) -> List[float]:
     if isinstance(node_size, (int, float)):
         return [node_size]
     elif isinstance(node_size, list):
@@ -123,11 +129,11 @@ def plot_plotly(
     mapper_plot,
     width: int,
     height: int,
-    node_size: Optional[Union[int, float, List[Union[int, float]]]] = DEFAULT_NODE_SIZE,
-    colors=None,
+    colors: Union[np.ndarray, List[float]],
+    node_size: Optional[Union[int, float, List[Union[int, float]]]] = None,
     title: Optional[Union[str, List[str]]] = None,
     agg=np.nanmean,
-    cmap: Union[str, List[str]] = DEFAULT_CMAP,
+    cmap: Optional[Union[str, List[str]]] = None,
 ) -> go.Figure:
     cmaps = _to_cmaps(cmap)
     colors = _to_colors(colors)
@@ -187,7 +193,7 @@ def plot_plotly_update(
     return fig
 
 
-def _node_pos_array(graph, dim, node_pos):
+def _node_pos_array(graph: nx.Graph, dim: int, node_pos):
     return tuple([node_pos[n][i] for n in graph.nodes()] for i in range(dim))
 
 
@@ -202,7 +208,7 @@ def _edge_pos_array(graph, dim, node_pos):
     return edges_arr
 
 
-def _marker_size(mapper_plot, node_size):
+def _marker_size(mapper_plot, node_size: float) -> List[float]:
     attr_size = nx.get_node_attributes(mapper_plot.graph, ATTR_SIZE)
     max_size = max(attr_size.values(), default=1.0)
[…]
