### 2025-06-02 665e3c5
Added app to src directory. Improved options with csv upload
