+++ src/tdamapper/app.py
@@ -1,9 +1,11 @@
+import logging
+
 import numpy as np
 import pandas as pd
 import plotly.graph_objs as go
 from nicegui import run, ui
 from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
-from sklearn.datasets import load_digits, load_iris
+from sklearn.datasets import fetch_openml, load_digits, load_iris
 from sklearn.decomposition import PCA
 from umap import UMAP
 
@@ -12,12 +14,23 @@ from tdamapper.cover import BallCover, CubicalCover, KNNCover
 from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
 
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
+
 
 def mode(arr):
     values, counts = np.unique(arr, return_counts=True)
     max_count = np.max(counts)
     mode_values = values[counts == max_count]
     return np.nanmean(mode_values)
+
+
+def _fix_data(data):
+    df = pd.DataFrame(data)
+    df = df.select_dtypes(include="number")
+    df.dropna(axis=1, how="all", inplace=True)
+    df.fillna(df.mean(), inplace=True)
+    return df
 
 
 def _identity(X):
@@ -66,6 +79,9 @@ DATA_SOURCE_EXAMPLE_IRIS = "Iris"
 DRAW_3D = "3D"
 DRAW_2D = "2D"
 DRAW_ITERATIONS = 50
+DRAW_MEAN = "Mean"
+DRAW_MEDIAN = "Median"
+DRAW_MODE = "Mode"
 
 
 class App:
@@ -96,7 +112,9 @@ class App:
             value=DATA_SOURCE_EXAMPLE,
         )
         self.data_source_csv = ui.upload(
-            on_upload=self.update_dataset_handler,
+            on_upload=self.update_csv_handler,
+            auto_upload=True,
+            label="Upload CSV",
         ).classes("w-full")
         self.data_source_csv.bind_visibility_from(
             target_object=self.data_source_type,
@@ -278,102 +296,166 @@ class App:
             value=DRAW_ITERATIONS,
             on_change=self.update_plot_handler,
         )
+        self.draw_aggregation = ui.select(
+            label="Aggregation",
+            options=[
+                DRAW_MEAN,
+                DRAW_MEDIAN,
+                DRAW_MODE,
+            ],
+            value=DRAW_MEAN,
+            on_change=self.update_plot_handler,
+        )
 
     def build_plot(self):
         fig = go.Figure()
         fig.layout.width = None
         fig.layout.autosize = True
         self.plot_container = ui.element("div").classes("w-full h-full")
-        with self.plot_container:
-            ui.plotly(go.Figure())
 
     def render_dataset(self):
         source_type = self.data_source_type.value
         if source_type == DATA_SOURCE_EXAMPLE:
             name = self.data_source_example_file.value
             if name == DATA_SOURCE_EXAMPLE_DIGITS:
-                X, y = load_digits(return_X_y=True, as_frame=True)
-                return X, y
+                df_X, df_y = load_digits(return_X_y=True, as_frame=True)
             elif name == DATA_SOURCE_EXAMPLE_IRIS:
-                X, y = load_iris(return_X_y=True, as_frame=True)
-                return X, y
+                df_X, df_y = load_iris(return_X_y=True, as_frame=True)
         elif source_type == DATA_SOURCE_CSV:
-            pass
+            csv_file = self.csv_file
+            if csv_file is None:
+                logger.warning("No CSV file uploaded")
+                df_X, df_y = pd.DataFrame(), pd.Series()
+            else:
+                df_X = pd.read_csv(csv_file.content)
[…]
