### 2025-06-02 146321a
Removed double leading underscore from internal attributes
