### 2025-06-02 dcbff54
Improved structure of plot backends
