### 2025-06-02 61a9ab2
Improved handling of UI elements in plotly plots. Removed deprecation
