+++ src/tdamapper/plot_backends/plot_plotly.py
@@ -53,29 +53,99 @@ def _get_plotly_colorscales():
 PLOTLY_CMAPS = _get_plotly_colorscales()
 
 
+class PlotlyUI:
+
+    def __init__(self):
+        self.menu_cmap = None
+        self.menu_color = None
+        self.slider_size = None
+
+    def set_menu_cmap(self, mapper_plot, cmaps: Optional[List[str]]) -> None:
+        if cmaps is None:
+            return
+        cmaps_plotly = [PLOTLY_CMAPS.get(c.lower()) for c in cmaps]
+        self.menu_cmap = _ui_cmap(mapper_plot, cmaps_plotly)
+
+    def set_menu_color(self, mapper_plot, colors, titles, agg):
+        self.menu_color = _ui_color(mapper_plot, colors, titles, agg)
+
+    def set_slider_size(self, mapper_plot, node_sizes):
+        self.slider_size = _ui_node_size(mapper_plot, node_sizes)
+
+
+def _to_cmaps(cmap: Optional[Union[str, List[str]]]) -> List[str]:
+    """Convert a single cmap or a list of cmaps to a list of cmaps."""
+    if cmap is None:
+        return [DEFAULT_CMAP]
+    if isinstance(cmap, str):
+        return [cmap]
+    elif isinstance(cmap, list):
+        return cmap
+    else:
+        raise ValueError(f"Invalid cmap type: {type(cmap)}. Expected str or list[str].")
+
+
+def _to_colors(colors: Union[np.ndarray, List[float]]) -> np.ndarray:
+    """Convert colors to a numpy array."""
+    colors_arr = np.array(colors)
+    if colors_arr.ndim == 1:
+        return colors_arr.reshape(-1, 1)
+    elif colors_arr.ndim == 2:
+        return colors_arr
+    else:
+        raise ValueError(
+            f"Invalid colors shape: {colors_arr.shape}. Expected 1D or 2D array."
+        )
+
+
+def _to_titles(title: Optional[Union[str, List[str]]], colors_num: int) -> List[str]:
+    if title is None:
+        return [f"{i}" for i in range(colors_num)]
+    elif isinstance(title, str):
+        return [f"{title} {i}" for i in range(colors_num)]
+    elif isinstance(title, list) and len(title) == colors_num:
+        return title
+    else:
+        raise ValueError(
+            f"Invalid title type: {type(title)}. Expected str or list[str]."
+        )
+
+
+def _to_node_sizes(
+    node_size: Optional[Union[int, float, List[Union[int, float]]]]
+) -> List[float]:
+    if isinstance(node_size, (int, float)):
+        return [node_size]
+    elif isinstance(node_size, list):
+        return node_size
+    else:
+        raise ValueError(
+            f"Invalid node_size type: {type(node_size)}. "
+            "Expected int, float or list[int, float]."
+        )
+
+
 def plot_plotly(
     mapper_plot,
     width: int,
     height: int,
-    node_size: Optional[Union[int, float, List[Union[int, float]]]] = DEFAULT_NODE_SIZE,
-    colors=None,
+    colors: Union[np.ndarray, List[float]],
+    node_size: Optional[Union[int, float, List[Union[int, float]]]] = None,
     title: Optional[Union[str, List[str]]] = None,
     agg=np.nanmean,
-    cmap: Union[str, List[str]] = DEFAULT_CMAP,
+    cmap: Optional[Union[str, List[str]]] = None,
 ) -> go.Figure:
-    cmaps = [cmap] if isinstance(cmap, str) else cmap
-    colors = np.array(colors)
[…]
