### 2025-06-02 7a5d85b
Improved notebooks settings
