+++ docs/source/notebooks/digits.py
@@ -103,10 +103,10 @@ fig = plot.plot_plotly(
     colors=labels,
     cmap=["jet", "viridis", "cividis"],
     agg=mode,
+    node_size=[0.0, 0.5, 1.0, 1.5, 2.0],
     title="mode of digits",
     width=600,
     height=600,
-    node_size=0.5,
 )
 
 fig.show(config={"scrollZoom": True}, renderer="notebook_connected")
@@ -134,10 +134,10 @@ fig = plot.plot_plotly(
     colors=labels,
     cmap=["jet", "viridis", "cividis"],
     agg=entropy,
+    node_size=[0.0, 0.5, 1.0, 1.5, 2.0],
     title="entropy of digits",
     width=600,
     height=600,
-    node_size=0.5,
 )
 
 fig.show(config={"scrollZoom": True}, renderer="notebook_connected")
