### 2025-06-02 67a2dc3
Refactored App with state and async callbacks
