+++ src/tdamapper/app.py
@@ -1,5 +1,7 @@
 import logging
+from dataclasses import asdict, dataclass
 
+import networkx as nx
 import numpy as np
 import pandas as pd
 import plotly.graph_objs as go
@@ -13,47 +15,6 @@ from tdamapper.core import TrivialClustering, TrivialCover
 from tdamapper.cover import BallCover, CubicalCover, KNNCover
 from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
-
-logger = logging.getLogger(__name__)
-logging.basicConfig(level=logging.INFO)
-
-
-def mode(arr):
-    values, counts = np.unique(arr, return_counts=True)
-    max_count = np.max(counts)
-    mode_values = values[counts == max_count]
-    return np.nanmean(mode_values)
-
-
-def _fix_data(data):
-    df = pd.DataFrame(data)
-    df = df.select_dtypes(include="number")
-    df.dropna(axis=1, how="all", inplace=True)
-    df.fillna(df.mean(), inplace=True)
-    return df
-
-
-def _identity(X):
-    return X
-
-
-def _pca(n_components):
-    pca = PCA(n_components=n_components, random_state=42)
-
-    def _func(X):
-        return pca.fit_transform(X)
-
-    return _func
-
-
-def _umap(n_components):
-    um = UMAP(n_components=n_components, random_state=42)
-
-    def _func(X):
-        return um.fit_transform(X)
-
-    return _func
-
 
 LENS_IDENTITY = "Identity"
 LENS_PCA = "PCA"
@@ -75,6 +36,7 @@ DATA_SOURCE_OPENML = "OpenML"
 
 DATA_SOURCE_EXAMPLE_DIGITS = "Digits"
 DATA_SOURCE_EXAMPLE_IRIS = "Iris"
+SOURCE_OPENML = "554"
 
 DRAW_3D = "3D"
 DRAW_2D = "2D"
@@ -82,6 +44,203 @@ DRAW_ITERATIONS = 50
 DRAW_MEAN = "Mean"
 DRAW_MEDIAN = "Median"
 DRAW_MODE = "Mode"
+
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
+
+
+@dataclass
+class State:
+    source_type: str = DATA_SOURCE_EXAMPLE
+    source_name: str = DATA_SOURCE_EXAMPLE_DIGITS
+    source_csv: str = "tmp/data.csv"
+    source_openml: str = SOURCE_OPENML
+    lens_type: str = LENS_PCA
+    lens_pca_n_components: int = 2
+    lens_umap_n_components: int = 2
+    cover_type: str = COVER_CUBICAL
+    cover_cubical_n_intervals: int = 2
+    cover_cubical_overlap_frac: float = 0.25
+    cover_knn_neighbors: int = 10
+    cover_ball_radius: float = 100.0
+    clustering_type: str = CLUSTERING_TRIVIAL
+    clustering_kmeans_n_clusters: int = 2
+    clustering_dbscan_eps: float = 0.5
+    clustering_dbscan_min_samples: int = 5
+    clustering_agglomerative_n_clusters: int = 2
+    draw_dim: int = 3
+    draw_aggregation: str = DRAW_MEAN
+    draw_iterations: int = 50
+
+
+def _fix_data(data):
+    df = pd.DataFrame(data)
+    df = df.select_dtypes(include="number")
+    df.dropna(axis=1, how="all", inplace=True)
+    df.fillna(df.mean(), inplace=True)
+    return df
+
+
+def get_dataset(state: State):
+    source_type = state.source_type
+    source_name = state.source_name
+    csv_file = state.source_csv
[…]
