+++ src/tdamapper/app.py
@@ -5,30 +5,41 @@ import networkx as nx
 import numpy as np
 import pandas as pd
 import plotly.graph_objs as go
-from nicegui import run, ui
+from nicegui import app, run, ui
 from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
 from sklearn.datasets import fetch_openml, load_digits, load_iris
 from sklearn.decomposition import PCA
 from umap import UMAP
 
-from tdamapper.core import TrivialClustering, TrivialCover
+from tdamapper.core import Cover, TrivialClustering, TrivialCover
 from tdamapper.cover import BallCover, CubicalCover, KNNCover
 from tdamapper.learn import MapperAlgorithm
 from tdamapper.plot import MapperPlot
 
+RANDOM_STATE = 42
 LENS_IDENTITY = "Identity"
 LENS_PCA = "PCA"
+LENS_PCA_N_COMPONENTS = 2
 LENS_UMAP = "UMAP"
+LENS_UMAP_N_COMPONENTS = 2
 
 COVER_TRIVIAL = "Trivial"
 COVER_CUBICAL = "Cubical"
+COVER_CUBICAL_N_INTERVALS = 2
+COVER_CUBICAL_OVERLAP_FRAC = 0.25
 COVER_BALL = "Ball"
+COVER_BALL_RADIUS = 100.0
 COVER_KNN = "KNN"
+COVER_KNN_NEIGHBORS = 10
 
 CLUSTERING_TRIVIAL = "Trivial"
 CLUSTERING_KMEANS = "KMeans"
+CLUSTERING_KMEANS_N_CLUSTERS = 2
 CLUSTERING_AGGLOMERATIVE = "Agglomerative"
+CLUSTERING_AGGLOMERATIVE_N_CLUSTERS = 2
 CLUSTERING_DBSCAN = "DBSCAN"
+CLUSTERING_DBSCAN_EPS = 0.5
+CLUSTERING_DBSCAN_MIN_SAMPLES = 5
 
 DATA_SOURCE_EXAMPLE = "Example"
 DATA_SOURCE_CSV = "CSV"
@@ -57,21 +68,21 @@ class State:
     source_csv: str = "tmp/data.csv"
     source_openml: str = SOURCE_OPENML
     lens_type: str = LENS_PCA
-    lens_pca_n_components: int = 2
-    lens_umap_n_components: int = 2
+    lens_pca_n_components: int = LENS_PCA_N_COMPONENTS
+    lens_umap_n_components: int = LENS_UMAP_N_COMPONENTS
     cover_type: str = COVER_CUBICAL
-    cover_cubical_n_intervals: int = 2
-    cover_cubical_overlap_frac: float = 0.25
-    cover_knn_neighbors: int = 10
-    cover_ball_radius: float = 100.0
+    cover_cubical_n_intervals: int = COVER_CUBICAL_N_INTERVALS
+    cover_cubical_overlap_frac: float = COVER_CUBICAL_OVERLAP_FRAC
+    cover_knn_neighbors: int = COVER_KNN_NEIGHBORS
+    cover_ball_radius: float = COVER_BALL_RADIUS
     clustering_type: str = CLUSTERING_TRIVIAL
-    clustering_kmeans_n_clusters: int = 2
-    clustering_dbscan_eps: float = 0.5
-    clustering_dbscan_min_samples: int = 5
-    clustering_agglomerative_n_clusters: int = 2
-    draw_dim: int = 3
+    clustering_kmeans_n_clusters: int = CLUSTERING_KMEANS_N_CLUSTERS
+    clustering_dbscan_eps: float = CLUSTERING_DBSCAN_EPS
[…]
