+++ src/tdamapper/app.py
@@ -65,7 +65,6 @@ logging.basicConfig(level=logging.INFO)
 class State:
     source_type: str = DATA_SOURCE_EXAMPLE
     source_name: str = DATA_SOURCE_EXAMPLE_DIGITS
-    source_csv: str = "tmp/data.csv"
     source_openml: str = SOURCE_OPENML
     lens_type: str = LENS_PCA
     lens_pca_n_components: int = LENS_PCA_N_COMPONENTS
@@ -91,38 +90,6 @@ def _fix_data(data):
     df.dropna(axis=1, how="all", inplace=True)
     df.fillna(df.mean(), inplace=True)
     return df
-
-
-def get_dataset(state: State):
-    source_type = state.source_type
-    source_name = state.source_name
-    csv_file = state.source_csv
-    openml_code = state.source_openml
-    df_X, df_y = pd.DataFrame(), pd.Series()
-    if source_type == DATA_SOURCE_EXAMPLE:
-        if source_name == DATA_SOURCE_EXAMPLE_DIGITS:
-            df_X, df_y = load_digits(return_X_y=True, as_frame=True)
-        elif source_name == DATA_SOURCE_EXAMPLE_IRIS:
-            df_X, df_y = load_iris(return_X_y=True, as_frame=True)
-    elif source_type == DATA_SOURCE_CSV:
-        if csv_file is None:
-            logger.warning("No CSV file uploaded")
-            df_X, df_y = pd.DataFrame(), pd.Series()
-        else:
-            df_X = pd.read_csv()
-            df_y = pd.Series()
-    elif source_type == DATA_SOURCE_OPENML:
-        if not openml_code:
-            logger.warning("No OpenML code provided")
-            df_X, df_y = pd.DataFrame(), pd.Series()
-        else:
-            df_X, df_y = fetch_openml(openml_code, return_X_y=True, as_frame=True)
-    else:
-        logger.error(f"Unknown data source type: {source_type}")
-        return pd.DataFrame(), pd.Series()
-    df_X = _fix_data(df_X)
-    df_y = _fix_data(df_y)
-    return df_X, df_y
 
 
 def get_lens(state: State):
@@ -191,18 +158,18 @@ def get_clustering(state: State):
         return TrivialClustering()
 
 
-def compute_mapper(**kwargs):
+def compute_mapper(df_X, labels, **kwargs):
     state = State(**kwargs)
 
-    df_X, labels = get_dataset(state)
+    # df_X, labels = get_dataset(state)
     if df_X.empty:
         logger.warning("No dataset loaded")
-        return None
+        return None, None
 
     lens = get_lens(state)
     if lens is None:
         logger.warning("No lens selected")
-        return None
+        return None, None
 
     X = df_X.to_numpy()
     y = lens(X)
@@ -210,12 +177,12 @@ def compute_mapper(**kwargs):
     cover = get_cover(state)
     if cover is None:
         logger.warning("No cover selected")
-        return None
+        return None, None
 
     clustering = get_clustering(state)
     if clustering is None:
         logger.warning("No clustering selected")
-        return None
+        return None, None
 
     mapper_algo = MapperAlgorithm(
         cover=cover,
@@ -260,13 +227,6 @@ def mode(arr):
     return np.nanmean(mode_values)
 
 
-def update_csv(file):
-    if file is None:
-        logger.warning("No file uploaded")
-        return None
-    return file
-
-
 class App:
 
     def build_dataset(self):
@@ -278,7 +238,7 @@ class App:
                 DATA_SOURCE_OPENML,
             ],
             value=DATA_SOURCE_EXAMPLE,
-            on_change=self.update,
+            on_change=self.load_dataset,
[…]
