### 2025-06-04 f288638
Improved state handling
