+++ src/tdamapper/app.py
@@ -1,5 +1,6 @@
 import logging
 from dataclasses import asdict, dataclass
+from typing import Any
 
 import networkx as nx
 import numpy as np
@@ -92,6 +93,38 @@ def _fix_data(data):
     return df
 
 
+def get_dataset(state: State, storage: dict[str, Any]):
+    source_type = state.source_type
+    source_name = state.source_name
+    csv_file = storage.get("csv_file", None)
+    openml_code = state.source_openml
+    df_X, df_y = pd.DataFrame(), pd.Series()
+    if source_type == DATA_SOURCE_EXAMPLE:
+        if source_name == DATA_SOURCE_EXAMPLE_DIGITS:
+            df_X, df_y = load_digits(return_X_y=True, as_frame=True)
+        elif source_name == DATA_SOURCE_EXAMPLE_IRIS:
+            df_X, df_y = load_iris(return_X_y=True, as_frame=True)
+    elif source_type == DATA_SOURCE_CSV:
+        if csv_file is None:
+            logger.warning("No CSV file uploaded")
+            df_X, df_y = pd.DataFrame(), pd.Series()
+        else:
+            df_X = pd.read_csv(csv_file)
+            df_y = pd.Series()
+    elif source_type == DATA_SOURCE_OPENML:
+        if not openml_code:
+            logger.warning("No OpenML code provided")
+            df_X, df_y = pd.DataFrame(), pd.Series()
+        else:
+            df_X, df_y = fetch_openml(openml_code, return_X_y=True, as_frame=True)
+    else:
+        logger.error(f"Unknown data source type: {source_type}")
+        return pd.DataFrame(), pd.Series()
+    df_X = _fix_data(df_X)
+    df_y = _fix_data(df_y)
+    return df_X, df_y
+
+
 def get_lens(state: State):
     def _pca(n):
         pca = PCA(n_components=n, random_state=RANDOM_STATE)
@@ -161,7 +194,6 @@ def get_clustering(state: State):
 def compute_mapper(df_X, labels, **kwargs):
     state = State(**kwargs)
 
-    # df_X, labels = get_dataset(state)
     if df_X.empty:
         logger.warning("No dataset loaded")
         return None, None
@@ -456,38 +488,6 @@ class App:
         fig.layout.autosize = True
         self.plot_container = ui.element("div").classes("w-full h-full")
 
-    def get_dataset(self):
-        state = self.state
-        source_type = state.source_type
-        source_name = state.source_name
-        csv_file = self.storage.get("csv_file", None)
-        openml_code = state.source_openml
-        df_X, df_y = pd.DataFrame(), pd.Series()
-        if source_type == DATA_SOURCE_EXAMPLE:
-            if source_name == DATA_SOURCE_EXAMPLE_DIGITS:
-                df_X, df_y = load_digits(return_X_y=True, as_frame=True)
-            elif source_name == DATA_SOURCE_EXAMPLE_IRIS:
-                df_X, df_y = load_iris(return_X_y=True, as_frame=True)
-        elif source_type == DATA_SOURCE_CSV:
-            if csv_file is None:
-                logger.warning("No CSV file uploaded")
-                df_X, df_y = pd.DataFrame(), pd.Series()
-            else:
-                df_X = pd.read_csv(csv_file)
-                df_y = pd.Series()
-        elif source_type == DATA_SOURCE_OPENML:
-            if not openml_code:
-                logger.warning("No OpenML code provided")
-                df_X, df_y = pd.DataFrame(), pd.Series()
-            else:
-                df_X, df_y = fetch_openml(openml_code, return_X_y=True, as_frame=True)
-        else:
[…]
