### 2025-06-04 06dabae
Added nicegui app
