+++ src/tdamapper/app.py
@@ -1,0 +1,390 @@
+import logging
+from dataclasses import asdict, dataclass
+
+import pandas as pd
+from nicegui import app, run, ui
+from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
+from sklearn.decomposition import PCA
+from umap import UMAP
+
+from tdamapper.clustering import TrivialClustering
+from tdamapper.cover import BallCover, CubicalCover, KNNCover
+from tdamapper.learn import MapperAlgorithm
+from tdamapper.plot import MapperPlot
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+LENS_IDENTITY = "Identity"
+LENS_PCA = "PCA"
+LENS_UMAP = "UMAP"
+
+LENS_PCA_N_COMPONENTS = 2
+LENS_UMAP_N_COMPONENTS = 2
+
+COVER_CUBICAL = "Cubical Cover"
+COVER_BALL = "Ball Cover"
+COVER_KNN = "KNN Cover"
+
+CLUSTERING_TRIVIAL = "Skip"
+CLUSTERING_KMEANS = "KMeans"
+CLUSTERING_DBSCAN = "DBSCAN"
+CLUSTERING_AGGLOMERATIVE = "Agglomerative Clustering"
+
+COVER_CUBICAL_N_INTERVALS = 10
+COVER_CUBICAL_OVERLAP_FRAC = 0.25
+COVER_KNN_NEIGHBORS = 10
+COVER_BALL_RADIUS = 100.0
+
+CLUSTERING_KMEANS_N_CLUSTERS = 2
+CLUSTERING_DBSCAN_EPS = 0.5
+CLUSTERING_DBSCAN_MIN_SAMPLES = 5
+CLUSTERING_AGGLOMERATIVE_N_CLUSTERS = 2
+
+RANDOM_SEED = 42
+
+
+@dataclass
+class MapperConfig:
+    lens_type: str = LENS_PCA
+    cover_type: str = COVER_CUBICAL
+    clustering_type: str = CLUSTERING_TRIVIAL
+    lens_pca_n_components: int = LENS_PCA_N_COMPONENTS
+    lens_umap_n_components: int = LENS_UMAP_N_COMPONENTS
+    cover_cubical_n_intervals: int = COVER_CUBICAL_N_INTERVALS
+    cover_cubical_overlap_frac: float = COVER_CUBICAL_OVERLAP_FRAC
+    cover_knn_neighbors: int = COVER_KNN_NEIGHBORS
+    clustering_kmeans_n_clusters: int = CLUSTERING_KMEANS_N_CLUSTERS
+    clustering_dbscan_eps: float = CLUSTERING_DBSCAN_EPS
+    clustering_dbscan_min_samples: int = CLUSTERING_DBSCAN_MIN_SAMPLES
+    clustering_agglomerative_n_clusters: int = CLUSTERING_AGGLOMERATIVE_N_CLUSTERS
+
+
+def identity(X):
+    return X
+
+
+def pca(n_components):
+
+    def _pca(X):
+        pca_model = PCA(n_components=n_components, random_state=RANDOM_SEED)
+        return pca_model.fit_transform(X)
+
+    return _pca
+
+
+def umap(n_components):
+
+    def _umap(X):
+        umap_model = umap.UMAP(n_components=n_components, random_state=RANDOM_SEED)
+        return umap_model.fit_transform(X)
+
+    return _umap
+
+
+def run_mapper(df, **kwargs):
+    if df is None:
+        logger.error("No data found. Please upload a file first.")
+        return
+    logger.info("Computing Mapper.")
+
[…]
