### 2025-06-05 2374ff9
Improved handling of UI elements
