+++ src/tdamapper/app.py
@@ -155,236 +155,264 @@ def run_mapper(df, **kwargs):
     ).plot_plotly(
         colors=X,
         height=800,
-        node_size=[0.0, 0.5, 1.0],
+        node_size=[i * 0.125 for i in range(17)],
     )
     logger.info("Mapper run completed successfully.")
     return mapper_fig
 
 
-@ui.page("/")
-def index():
-    storage = app.storage.client
+class App:
 
-    def upload_file(file):
+    def __init__(self, storage):
+        self.storage = storage
+        ui.query("body").style("margin: 0; padding: 0; overflow: hidden;")
+        with ui.row().classes("w-full h-screen m-0 p-0 gap-0 overflow-hidden"):
+            with ui.column().classes("w-64 h-full m-0 p-0 overflow-y-auto"):
+                self._init_file_upload()
+                self._init_lens()
+                self._init_cover()
+                self._init_clustering()
+
+                ui.button(
+                    "Run Mapper",
+                    on_click=self.async_run_mapper,
+                    color="primary",
+                ).classes("w-full")
+
+            with ui.column().classes("flex-grow h-full overflow-hidden p-0 m-0"):
+                self._init_plot()
+
+    def _init_file_upload(self):
+        with ui.card().tight().classes("w-full p-3"):
+            ui.upload(
+                on_upload=self.upload_file,
+                auto_upload=True,
+                label="Upload CSV File",
+            ).classes("w-full")
+            with ui.card_section().classes("w-full"):
+                ui.button("Load", on_click=self.load_file).classes("w-full")
+
+    def _init_lens(self):
+        with ui.card().classes("w-full p-3"):
+            self.lens_type = ui.select(
+                options=[
+                    LENS_IDENTITY,
+                    LENS_PCA,
+                    LENS_UMAP,
+                ],
+                label="Lens",
+                value=LENS_PCA,
+            ).classes("w-full")
+
+            self.lens_pca_n_components = ui.number(
+                label="PCA Components",
+                value=LENS_PCA_N_COMPONENTS,
+            ).classes("w-full")
+            self.lens_pca_n_components.bind_visibility_from(
+                target_object=self.lens_type,
+                target_name="value",
+                value=LENS_PCA,
+            )
+
+            self.lens_umap_n_components = ui.number(
+                label="UMAP Components",
+                value=LENS_UMAP_N_COMPONENTS,
+            ).classes("w-full")
+            self.lens_umap_n_components.bind_visibility_from(
+                target_object=self.lens_type,
+                target_name="value",
+                value=LENS_UMAP,
+            )
+
+    def _init_cover(self):
+        with ui.card().classes("w-full p-3"):
+            self.cover_type = ui.select(
+                options=[
+                    COVER_CUBICAL,
+                    COVER_BALL,
+                    COVER_KNN,
+                ],
+                label="Cover",
+                value=COVER_CUBICAL,
+            ).classes("w-full")
+
+            self.cover_cubical_n_intervals = ui.number(
+                label="Number of Intervals",
+                value=COVER_CUBICAL_N_INTERVALS,
+            ).classes("w-full")
+            self.cover_cubical_n_intervals.bind_visibility_from(
+                target_object=self.cover_type,
+                target_name="value",
+                value=COVER_CUBICAL,
+            )
+
+            self.cover_cubical_overlap_frac = ui.number(
+                label="Ball Radius",
+                value=COVER_BALL_RADIUS,
+            ).classes("w-full")
+            self.cover_cubical_overlap_frac.bind_visibility_from(
+                target_object=self.cover_type,
+                target_name="value",
+                value=COVER_BALL,
+            )
+
+            self.cover_knn_neighbors = ui.number(
[…]
