### 2025-06-05 ca4c2f2
Added scaling toggle. UI improvements
