+++ src/tdamapper/app.py
@@ -5,6 +5,7 @@ import pandas as pd
 from nicegui import app, run, ui
 from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
 from sklearn.decomposition import PCA
+from sklearn.preprocessing import StandardScaler
 from umap import UMAP
 
 from tdamapper.clustering import TrivialClustering
@@ -23,6 +24,8 @@ LENS_UMAP = "UMAP"
 LENS_PCA_N_COMPONENTS = 2
 LENS_UMAP_N_COMPONENTS = 2
 
+COVER_SCALE_DATA = False
+
 COVER_CUBICAL = "Cubical Cover"
 COVER_BALL = "Ball Cover"
 COVER_KNN = "KNN Cover"
@@ -37,6 +40,8 @@ COVER_CUBICAL_OVERLAP_FRAC = 0.25
 COVER_KNN_NEIGHBORS = 10
 COVER_BALL_RADIUS = 100.0
 
+CLUSTERING_SCALE_DATA = False
+
 CLUSTERING_KMEANS_N_CLUSTERS = 2
 CLUSTERING_DBSCAN_EPS = 0.5
 CLUSTERING_DBSCAN_MIN_SAMPLES = 5
@@ -48,12 +53,15 @@ RANDOM_SEED = 42
 @dataclass
 class MapperConfig:
     lens_type: str = LENS_PCA
+    cover_scale_data: bool = COVER_SCALE_DATA
     cover_type: str = COVER_CUBICAL
+    clustering_scale_data: bool = CLUSTERING_SCALE_DATA
     clustering_type: str = CLUSTERING_TRIVIAL
     lens_pca_n_components: int = LENS_PCA_N_COMPONENTS
     lens_umap_n_components: int = LENS_UMAP_N_COMPONENTS
     cover_cubical_n_intervals: int = COVER_CUBICAL_N_INTERVALS
     cover_cubical_overlap_frac: float = COVER_CUBICAL_OVERLAP_FRAC
+    cover_ball_radius: float = COVER_BALL_RADIUS
     cover_knn_neighbors: int = COVER_KNN_NEIGHBORS
     clustering_kmeans_n_clusters: int = CLUSTERING_KMEANS_N_CLUSTERS
     clustering_dbscan_eps: float = CLUSTERING_DBSCAN_EPS
@@ -61,11 +69,19 @@ class MapperConfig:
     clustering_agglomerative_n_clusters: int = CLUSTERING_AGGLOMERATIVE_N_CLUSTERS
 
 
-def identity(X):
+def fix_data(data):
+    df = pd.DataFrame(data)
+    df = df.select_dtypes(include="number")
+    df.dropna(axis=1, how="all", inplace=True)
+    df.fillna(df.mean(), inplace=True)
+    return df
+
+
+def lens_identity(X):
     return X
 
 
-def pca(n_components):
+def lens_pca(n_components):
 
     def _pca(X):
         pca_model = PCA(n_components=n_components, random_state=RANDOM_SEED)
@@ -74,11 +90,11 @@ def pca(n_components):
     return _pca
 
 
-def umap(n_components):
+def lens_umap(n_components):
 
     def _umap(X):
-        umap_model = umap.UMAP(n_components=n_components, random_state=RANDOM_SEED)
-        return umap_model.fit_transform(X)
+        um = UMAP(n_components=n_components, random_state=RANDOM_SEED)
+        return um.fit_transform(X)
 
     return _umap
 
@@ -92,12 +108,15 @@ def run_mapper(df, **kwargs):
     mapper_config = MapperConfig(**kwargs)
 
     lens_type = mapper_config.lens_type
+    cover_scale_data = mapper_config.cover_scale_data
     cover_type = mapper_config.cover_type
[…]
