### 2025-06-07 c7fc1a5
Added data examples. Improved UI
