+++ src/tdamapper/app.py
@@ -4,6 +4,7 @@ from dataclasses import asdict, dataclass
 import pandas as pd
 from nicegui import app, run, ui
 from sklearn.cluster import DBSCAN, AgglomerativeClustering, KMeans
+from sklearn.datasets import load_digits, load_iris
 from sklearn.decomposition import PCA
 from sklearn.preprocessing import StandardScaler
 from umap import UMAP
@@ -23,6 +24,11 @@ ICON_URL = f"{GIT_REPO_URL}/raw/main/docs/source/logos/tda-mapper-logo-icon.png"
 
 LOGO_URL = f"{GIT_REPO_URL}/raw/main/docs/source/logos/tda-mapper-logo-horizontal.png"
 
+
+LOAD_EXAMPLE = "Example"
+LOAD_EXAMPLE_DIGITS = "Digits"
+LOAD_EXAMPLE_IRIS = "Iris"
+LOAD_CSV = "CSV"
 
 LENS_IDENTITY = "Identity"
 LENS_PCA = "PCA"
@@ -106,8 +112,8 @@ def lens_umap(n_components):
     return _umap
 
 
-def run_mapper(df, **kwargs):
-    if df is None:
+def run_mapper(df, labels, **kwargs):
+    if df is None or df.empty:
         logger.error("No data found. Please upload a file first.")
         return
     logger.info("Computing Mapper.")
@@ -133,12 +139,14 @@ def run_mapper(df, **kwargs):
     )
 
     lens = lens_pca(n_components=LENS_PCA_N_COMPONENTS)
+    lens_name = LENS_PCA
     if lens_type == LENS_IDENTITY:
         lens = lens_identity
     elif lens_type == LENS_PCA:
         lens = lens_pca(n_components=lens_pca_n_components)
     elif lens_type == LENS_UMAP:
         lens = lens_umap(n_components=lens_umap_n_components)
+    lens_name = lens_type
 
     if cover_type == COVER_CUBICAL:
         cover = CubicalCover(
@@ -177,16 +185,20 @@ def run_mapper(df, **kwargs):
     df_fixed = fix_data(df)
     X = df_fixed.to_numpy()
     y = lens(X)
+    df_y = pd.DataFrame(y, columns=[f"{lens_type} {i}" for i in range(y.shape[1])])
+    df_labels = pd.DataFrame(labels) if labels is not None else pd.DataFrame()
     if cover_scale_data:
         y = StandardScaler().fit_transform(y)
     if clustering_scale_data:
         X = StandardScaler().fit_transform(X)
+    df_colors = pd.concat([df_labels, df_y, df_fixed], axis=1)
     mapper_graph = mapper.fit_transform(X, y)
     mapper_fig = MapperPlot(
         mapper_graph,
         dim=3,
     ).plot_plotly(
-        colors=X,
+        colors=df_colors.to_numpy(),
+        title=df_colors.columns.to_list(),
         height=800,
         node_size=[i * 0.125 for i in range(17)],
     )
@@ -198,41 +210,80 @@ class App:
 
     def __init__(self, storage):
         self.storage = storage
-        with ui.row().classes("w-full h-screen overflow-hidden p-0 m-0"):
-            with ui.column().classes("w-96 h-full p-1 m-0"):
-                with ui.link(target=GIT_REPO_URL, new_tab=True).classes(
-                    "w-full p-1 m-0"
-                ):
-                    ui.image(LOGO_URL)
-                with ui.column().classes("w-full h-full overflow-y-auto p-1 m-0"):
-                    self._init_file_upload()
-                    self._init_lens()
-                    self._init_cover()
-                    self._init_clustering()
+        with ui.left_drawer(elevated=True).classes(
+            "w-96 h-full overflow-y-auto gap-12"
+        ):
[…]
