### 2025-06-07 b1e907f
Fixed position of plotly menus in case of empty buttons
