### 2025-06-07 48e7940
Added colorscales. Added script
