### 2025-06-08 ac3fb5d
Added port configuration from environment variable
