### 2025-06-08 1ed83be
Added host from env variable
