### 2025-06-08 2b1c651
Fixed readme and docs. Added redirect
