### 2025-06-09 #222 @yeelauren
## Editor response to review:

👋 Hi @lucasimi we have reviewers ! 🥳  Thanks so much for your patience!
---

## Editor comments
:wave: Hi  @sampottinger and  @elliesch! Thank you for volunteering to review
for pyOpenSci!

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due:  June 30th, 2025
---

Reviewers: @sampottinger and  @elliesch
Due date:June 30th, 2025
