### 2025-06-09 9f521ff
Made app module internal. Added unit test for app
