### 2025-06-09 2a7175c
added condition to skip app testing
