+++ .github/workflows/test-bench.yml
@@ -49,7 +49,7 @@ jobs:
 
       - name: Install dependencies
         run: | 
-          python -m pip install -e .[dev]
+          python -m pip install -e .[dev, app]
 
       - name: Run benchmarks
         run: | 
