+++ .github/workflows/test-unit.yml
@@ -49,7 +49,7 @@ jobs:
 
       - name: Install dependencies
         run: | 
-          python -m pip install -e .[dev]
+          python -m pip install -e .[dev, app]
 
       - name: Run tests and code coverage
         run: | 
