+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,0 +1,42 @@
+---
+name: Bug Report
+about: Create a report to help us improve
+title: "[Bug]"
+labels: ''
+assignees: ''
+
+---
+
+## 🚨 Description
+
+Describe the bug clearly and concisely. What did you expect to happen, and what actually happened?
+
+## 🤖 Steps to Reproduce
+
+List the exact steps to reproduce the issue. For example:
+
+1. Go to '...'
+2. Click on '...'
+3. Scroll down to '...'
+4. See the error
+
+## 🧩 Expected Behavior
+
+What *should* happen instead?
+
+## 🖥️ Environment
+
+Please provide details about your environment:
+
+- OS:
+- Python version:
+- Library version:
+- Browser (if applicable):
+
+## 📸 Screenshots / Logs (Optional)
+
+Add screenshots or paste error logs here if available.
+
+## ✅ Additional Context (Optional)
+
+Any other relevant context or information about the problem?
