+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -1,0 +1,25 @@
+---
+name: Feature Request
+about: Suggest an idea for this project
+title: "[Request]"
+labels: ''
+assignees: ''
+
+---
+
+## 🚀 Feature Request
+
+Describe the feature you'd like. Clearly explain what you want to add or improve.
+
+## 🔍 Motivation
+
+Why is this feature important or useful? Explain what problem it solves or how it improves the user/developer experience.
+
+## 🧠 Algorithm Details (Optional)
+
+Is this related to an algorithm in the package? If so, please provide technical details, such as:
+
+- Which algorithm this affects or extends
+- Suggested changes or improvements
+- Any relevant theoretical background
+- (Optional) Benchmarks or comparisons to support your proposal
