### 2025-06-10 #222 @lucasimi
Hi! Thank you so much in advance for taking the time to look into this. I really value your input and I think your feedback will make a big difference in improving the overall quality of the package. @yeelauren @sampottinger @aflip
