### 2025-06-12 4b3cda0
Fixed edge colors on first rendering. Improved default sizes
