+++ src/tdamapper/plot_backends/plot_plotly.py
@@ -4,7 +4,7 @@ plotly.
 """
 
 import math
-from typing import List, Optional, Union
+from typing import Any, List, Optional, Union
 
 import networkx as nx
 import numpy as np
@@ -127,20 +127,29 @@ def _to_node_sizes(
 
 def plot_plotly(
     mapper_plot,
-    width: int,
-    height: int,
     colors: Union[np.ndarray, List[float]],
     node_size: Optional[Union[int, float, List[Union[int, float]]]] = None,
     title: Optional[Union[str, List[str]]] = None,
     agg=np.nanmean,
     cmap: Optional[Union[str, List[str]]] = None,
+    width: Optional[int] = None,
+    height: Optional[int] = None,
 ) -> go.Figure:
     cmaps = _to_cmaps(cmap)
     colors = _to_colors(colors)
     colors_num = colors.shape[1]
     titles = _to_titles(title, colors_num)
     node_sizes = _to_node_sizes(node_size)
-    fig = _figure(mapper_plot, width, height, node_sizes, colors, titles, agg, cmaps)
+    fig = _figure(
+        mapper_plot=mapper_plot,
+        node_sizes=node_sizes,
+        colors=colors,
+        titles=titles,
+        agg=agg,
+        cmaps=cmaps,
+        width=width,
+        height=height,
+    )
     ui = PlotlyUI()
     ui.set_menu_cmap(mapper_plot, cmaps)
     ui.set_menu_color(mapper_plot, colors, titles, agg)
@@ -250,35 +259,44 @@ def _set_cmap(mapper_plot, fig: go.Figure, cmap: str) -> None:
         )
 
 
+def _edge_colors_from_node_colors(
+    graph: nx.Graph,
+    node_colors: dict[Any, float],
+) -> List[float]:
+    edge_col = []
+    for e in graph.edges():
+        c0, c1 = node_colors[e[0]], node_colors[e[1]]
+        edge_col.append(c0)
+        edge_col.append(c1)
+        edge_col.append(c1)
+    return edge_col
+
+
 def _set_colors(mapper_plot, fig: go.Figure, colors, agg):
-    node_col = aggregate_graph(colors, mapper_plot.graph, agg)
-    scatter_text = _text(mapper_plot, node_col)
-    colors_arr = list(node_col.values())
+    node_col_agg = aggregate_graph(colors, mapper_plot.graph, agg)
+    node_col_arr = list(node_col_agg.values())
+    scatter_text = _text(mapper_plot, node_col_agg)
     fig.update_traces(
         patch=dict(
             text=scatter_text,
             marker=dict(
-                color=colors_arr,
-                cmin=min(colors_arr, default=None),
-                cmax=max(colors_arr, default=None),
+                color=node_col_arr,
+                cmin=min(node_col_arr, default=None),
+                cmax=max(node_col_arr, default=None),
             ),
         ),
         selector=dict(name=_NODES_TRACE),
     )
     if mapper_plot.dim == 3:
-        colors_avg = []
-        for e in mapper_plot.graph.edges():
-            c0, c1 = node_col[e[0]], node_col[e[1]]
-            colors_avg.append(c0)
-            colors_avg.append(c1)
-            colors_avg.append(c1)
+        edge_col = _edge_colors_from_node_colors(
+            mapper_plot.graph,
+            node_col_agg,
+        )
         fig.update_traces(
             patch=dict(
-                marker=dict(
-                    line_color=colors_avg,
-                    line_cmin=min(colors_arr, default=None),
-                    line_cmax=max(colors_arr, default=None),
-                ),
+                line_color=edge_col,
+                line_cmin=min(node_col_arr, default=None),
+                line_cmax=max(node_col_arr, default=None),
             ),
             selector=dict(name=_EDGES_TRACE),
         )
[…]
