### 2025-06-12 c688996
Fixed vanishing colormap title
