### 2025-06-13 bf11883
Applied class based approach to plotly plot
