+++ docs/source/notebooks/digits.py
@@ -103,7 +103,7 @@ fig = plot.plot_plotly(
     colors=labels,
     cmap=["jet", "viridis", "cividis"],
     agg=mode,
-    node_size=[0.0, 0.5, 1.0, 1.5, 2.0],
+    node_size=[0.25 * x for x in range(9)],
     title="mode of digits",
 )
 
@@ -132,7 +132,7 @@ fig = plot.plot_plotly(
     colors=labels,
     cmap=["jet", "viridis", "cividis"],
     agg=entropy,
-    node_size=[0.0, 0.5, 1.0, 1.5, 2.0],
+    node_size=[0.25 * x for x in range(9)],
     title="entropy of digits",
 )
 
