+++ src/tdamapper/plot_backends/plot_plotly.py
@@ -31,6 +31,8 @@ _NODES_TRACE = "nodes_trace"
 
 _EDGES_TRACE = "edges_trace"
 
+_DEFAULT_SPACING = 0.25
+
 DEFAULT_NODE_SIZE = 1
 
 DEFAULT_CMAP = "Jet"
@@ -38,7 +40,7 @@ DEFAULT_CMAP = "Jet"
 DEFAULT_TITLE = ""
 
 
-def _get_plotly_colorscales():
+def _get_plotly_colorscales() -> Dict[str, str]:
     modules = ["sequential", "diverging", "cyclical"]
     all_scales = []
     for mod in modules:
@@ -53,24 +55,9 @@ def _get_plotly_colorscales():
 PLOTLY_CMAPS = _get_plotly_colorscales()
 
 
-class PlotlyUI:
-
-    def __init__(self):
-        self.menu_cmap = None
-        self.menu_color = None
-        self.slider_size = None
-
-    def set_menu_cmap(self, mapper_plot, cmaps: Optional[List[str]]) -> None:
-        if cmaps is None:
-            return
-        cmaps_plotly = [PLOTLY_CMAPS.get(c.lower()) for c in cmaps]
-        self.menu_cmap = _ui_cmap(mapper_plot, cmaps_plotly)
-
-    def set_menu_color(self, mapper_plot, colors, titles, agg):
-        self.menu_color = _ui_color(mapper_plot, colors, titles, agg)
-
-    def set_slider_size(self, mapper_plot, node_sizes):
-        self.slider_size = _ui_node_size(mapper_plot, node_sizes)
+def _fmt(x: Any, max_len: int = 3) -> str:
+    fmt = f".{max_len}g"
+    return f"{x:{fmt}}"
 
 
 def _to_cmaps(cmap: Optional[Union[str, List[str]]]) -> List[str]:
@@ -102,7 +89,10 @@ def _to_titles(title: Optional[Union[str, List[str]]], colors_num: int) -> List[
     if title is None:
         return [f"{i}" for i in range(colors_num)]
     elif isinstance(title, str):
-        return [f"{title} {i}" for i in range(colors_num)]
+        if colors_num == 1:
+            return [title]
+        else:
+            return [f"{title} [{i}]" for i in range(colors_num)]
     elif isinstance(title, list) and len(title) == colors_num:
         return title
     else:
@@ -125,6 +115,13 @@ def _to_node_sizes(
         )
 
 
+def _get_cmap_rgb(cmap: str):
+    """Return a colorscale in [[float, 'rgb(r,g,b)']] format."""
+    base_scale = pc.get_colorscale(cmap)
+    # If it's already in [float, color] format, we're good
+    return [[pos, color] for pos, color in base_scale]
+
+
 def plot_plotly(
     mapper_plot,
     colors: Union[np.ndarray, List[float]],
@@ -140,21 +137,16 @@ def plot_plotly(
     colors_num = colors.shape[1]
     titles = _to_titles(title, colors_num)
     node_sizes = _to_node_sizes(node_size)
-    fig = _figure(
-        mapper_plot=mapper_plot,
-        node_sizes=node_sizes,
+    plot = PlotlyPlot(mapper_plot)
+    fig = plot.plot(
         colors=colors,
+        node_sizes=node_sizes,
         titles=titles,
         agg=agg,
         cmaps=cmaps,
         width=width,
         height=height,
     )
-    ui = PlotlyUI()
-    ui.set_menu_cmap(mapper_plot, cmaps)
-    ui.set_menu_color(mapper_plot, colors, titles, agg)
[…]
