### 2025-07-01 #222 @sampottinger
Hello! Thank you so much @lucasimi for this work and your patience. I want to congratulate you on a very readable codebase with strong API documentation and excellent examples. I have my first draft of my review below for your consideration. Please let me know what you think and I am very happy to chat further.

- [ ] ~~Any additional setup required to use the package (authentication tokens, etc.)~~

Omitted as the submission is not also going to JOSS afterwards.

---

Thank you for an excellent package. I think your work is usable, the documentation is clear, the implementation is strong, and the contribution is important. I have some suggestions and I have:

 - Opened issues for the 3 recommendations I made that I think may be best to address or further discuss prior to acceptance.
 - I have also flagged 2 additional comment for you to discuss below prior to final approval but I'm uncertain if perhaps the current behavior is intended so did not open an issue.
 - I have included some optional comments for you to consider but certainly should not hold up acceptance.

All that said, I would love to discuss this feedback with you and see if there are any other ways I may be helpful.  Thank again for your contributions to the community. I'm very happy to glad to have this library as part of the ecosystem!

**Readme and documentation**

The README is very well written and the documentation is extensive thanks to strong inline docstrings across the project. Even so, it may require some expansion to conform to pyOpenSci guidance.

Comments

 - Installation was a breeze.
 - I want to applaud the choice to keep minimal requirements short with the separate set of app dependencies.
 - The use of the circles dataset as something familiar to immediately ground the work was great.

Recommended

 - I love your documentation but I don't think you have examples for all user facing classes and functions. In particular, I think you might need examples for the other cover options you provide like `BallCover`. Just some commentary on what the options are and / or when one might want to use it would be fantastic. See https://github.com/lucasimi/tda-mapper-python/issues/234.
 - You need a https://www.repostatus.org/ to conform with pyopensci guidance. See https://github.com/lucasimi/tda-mapper-python/issues/235.

Consider (optional)

- You may consider adding a status indicator to classifiers such as `Development Status :: 5 - Production/Stable` to your `pyproject.toml`.
- I wonder if the quick start (and installation) may come before background and citations. Those using this library likely already know about the mapper algorithm so cutting to the quick start might be useful.

**Usage**

I find the API intuitive. The design is familiar and easy to engage with after the documentation.

Comments

 - I did a few checks with the examples as starting points to try variations such as other cover algorithms. I found it easy to use your documentation to explore other options.
 - The visualization functionality and scikit-learn integration are excellent.
 - `_common.py` is doing a good job on validation.

Recommended

 - Tests did pass according to contributor instructions. Even so,  you should add `nicegui` and `umap-learn` to dev in `pyproject.toml` or describe the correct local setup in either `CONTRIBUTING.md` or `README.md` as you do ask developers to ensure tests pass. See https://github.com/lucasimi/tda-mapper-python/issues/236.

Consider (optional)

 - If this library is relatively new, I wonder if you may consider removing `tdamapper.core.MapperAlgorithm` and other deprecated classes. However, this is low risk given the very clear deprecation annotation.
 - You may consider adding a docstring in `__init__.py` for those using `help` commands with very high level library usage guidance. This could be very short with a minimal description and example.
 - I love that you included a [live demo](https://tda-mapper-app.up.railway.app/). I don't have an immediate solution for this but it was showing blank for me when I landed on the page until I made the window bigger.

**Implementation**

Comments

 - Code docstring coverage in production code is good, especially in `core.py`.
 - I was unclear what was going on in `main` but one final look before submitting and it appears [you may be failing CI / CD](https://github.com/lucasimi/tda-mapper-python/actions/runs/15959545925).

Recommend

 - Unless there is a specific reason, you should consider removing 3.8 in `pyproject.toml` as it is EOL. *However, I have not opened an issue for this because I am unclear if maybe 3.8 was included past EOL for some specific reason.*
 - The app itself (`app.py`, `streamlit_app.py`) may benefit from inline documentation. *However, I have not opened an issue to have discussion here if you feel strongly.*

Consider (optional)

 - I would still suggest adding at least short descriptions to supporting structures with underscore prefix like to `Pullback` and `MapperAlgorithm`. However, these are not public members so this is optional. However, even just a description of their purpose would be helpful.
 - The tests are simple but coverage is broad. My only concern is that the the tests are not very varied and test data are not well described (like X in `test_unit_knn.py`).  Some comments here may help.
 - You have type information in a number of docstrings but you don't use actual type hints. For  the places where this is available, this would help mypy users.
 - I know its fairly straightforward but you should consider adding a docstring to `EstimatorMixin` similar to `ParamsMixin`. In general, public members should have documentation even if just for other developers.
