### 2025-07-04 #222 @lucasimi
Hi @sampottinger 👋, thank you so much for your thoughtful review and your kind words!
I think your points are valid and the improvements you suggest would greatly help the users of the package.
In the following days I'll start working on them, and go back to you if needed.

For now I'll briefly go through the points of your review.

1. Readme and documentation: I think expanding the examples makes a lot sense as you suggest. This is important because many user-facing functions require some context to be easily understood. Regarding the sectioning of the README, unfortunately the name "mapper" is over-used in the tech context: if you run a search on github you can see that many results come from very far contexts (for example object mapping in relational databases). For this reason I decided to go straight to the TDA context in the first part of the README.

2. Usage: I agree, those requirements should be improved a bit as you suggest. I agree also on the deprecations, I was planning to remove them at some point. Regarding the app, the left sidebar is collapsed when the width is lower than a threshold. This behavior is hard-coded in `nicegui`, but I think there are ways around that. The user can still drag the leftmost pixel in the browser window, and the left drawer is visible again. However this is not a solution. A better improvement would be to add a top header with a toggle button, or something in that direction.

3. Implementation: I agree, Python 3.8 could be removed, there is no strong need to have it. I have a plan to add typing everywhere, since it improves dev experience and user experience by a lot. [The link you pasted](https://github.com/lucasimi/tda-mapper-python/actions/runs/15959545925) is not from _main_ but from a development branch, so it's expected to fail for now since it's work in progress. Regarding CI however, I think that codecov is configured in a way that CI fails too easily, and this could be improved too. Regarding tests, I agree that the datasets could be improved. Simple tests are still needed because it's easy to reason about them when you know the definition of the cover algorithms you want to test.
