### 2025-07-19 #222 @lucasimi
Hi @sampottinger! I finally made some changes to the code targeting your review and I'm ready to give you a more details feedback.

> #### Review Comments
>
> Thank you for an excellent package. I think your work is usable, the documentation is clear, the implementation is strong, and the contribution is important. I have some suggestions and I have:
>
>     * Opened issues for the 3 recommendations I made that I think may be best to address or further discuss prior to acceptance.
>
>     * I have also flagged 2 additional comment for you to discuss below prior to final approval but I'm uncertain if perhaps the current behavior is intended so did not open an issue.
>
>     * I have included some optional comments for you to consider but certainly should not hold up acceptance.
>

All the issues have been solved and their PRs have been merged to `main`. I also worked on the other suggestions and comments you provided. See below for more info.

>
> All that said, I would love to discuss this feedback with you and see if there are any other ways I may be helpful. Thank again for your contributions to the community. I'm very happy to glad to have this library as part of the ecosystem!
>
> **Readme and documentation**
>
> The README is very well written and the documentation is extensive thanks to strong inline docstrings across the project. Even so, it may require some expansion to conform to pyOpenSci guidance.
>
> Comments
>
>     * Installation was a breeze.
>
>     * I want to applaud the choice to keep minimal requirements short with the separate set of app dependencies.
>
>     * The use of the circles dataset as something familiar to immediately ground the work was great.
>
>
> Recommended
>
>     * I love your documentation but I don't think you have examples for all user facing classes and functions. In particular, I think you might need examples for the other cover options you provide like `BallCover`. Just some commentary on what the options are and / or when one might want to use it would be fantastic. See [[Request] Expand documentation on other cover options lucasimi/tda-mapper-python#234](https://github.com/lucasimi/tda-mapper-python/issues/234).
>
>     * You need a https://www.repostatus.org/ to conform with pyopensci guidance. See [[General] Add repo / development status lucasimi/tda-mapper-python#235](https://github.com/lucasimi/tda-mapper-python/issues/235).
>

A detailed explanation about the cover algorithms has been added as an example [in the documentation](https://tda-mapper.readthedocs.io/en/main/notebooks/cover.html). The _repostatus_ badge has been added in the [README](https://github.com/lucasimi/tda-mapper-python/blob/main/README.md) and in the [documentation home page](https://tda-mapper.readthedocs.io/en/main/index.html).

>
> Consider (optional)
>
>     * You may consider adding a status indicator to classifiers such as `Development Status :: 5 - Production/Stable` to your `pyproject.toml`.
>

It's been added [here](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/pyproject.toml#L29).

>     * I wonder if the quick start (and installation) may come before background and citations. Those using this library likely already know about the mapper algorithm so cutting to the quick start might be useful.
>

I decided to move the citations in the last part of the [README](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/pyproject.toml#L29) and [in the docs](https://tda-mapper.readthedocs.io/en/main/citations.html). However, I'm not decided yet on weather moving the background section after the quickstart.

>
> **Usage**
>
> I find the API intuitive. The design is familiar and easy to engage with after the documentation.
>
> Comments
>
>     * I did a few checks with the examples as starting points to try variations such as other cover algorithms. I found it easy to use your documentation to explore other options.
>
>     * The visualization functionality and scikit-learn integration are excellent.
>
>     * `_common.py` is doing a good job on validation.
>
>
> Recommended
>
>     * Tests did pass according to contributor instructions. Even so,  you should add `nicegui` and `umap-learn` to dev in `pyproject.toml` or describe the correct local setup in either `CONTRIBUTING.md` or `README.md` as you do ask developers to ensure tests pass. See [[General] Expand the dev requirements lucasimi/tda-mapper-python#236](https://github.com/lucasimi/tda-mapper-python/issues/236).
>

Missing dependencies have been added [here](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/pyproject.toml#L56).

>
> Consider (optional)
>
>     * If this library is relatively new, I wonder if you may consider removing `tdamapper.core.MapperAlgorithm` and other deprecated classes. However, this is low risk given the very clear deprecation annotation.
>

I agree, all the deprecated functionalities have been removed. There is no point in keeping them any longer.

>     * You may consider adding a docstring in `__init__.py` for those using `help` commands with very high level library usage guidance. This could be very short with a minimal description and example.
>

I've added a very general overview of the package [in this file](https://github.com/lucasimi/tda-mapper-python/blob/main/src/tdamapper/__init__.py). I'm not sure if this would be enough.

>     * I love that you included a [live demo](https://tda-mapper-app.up.railway.app/). I don't have an immediate solution for this but it was showing blank for me when I landed on the page until I made the window bigger.
>

The issue [has been solved](https://tda-mapper-app.up.railway.app) by adding a [menu button](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/src/tdamapper/app.py#L704), which is visible only when the left drawer is closed.
FYI: now the app is taking a little longer to boot because I'm saving money on deployment: I opted for _sleep mode_ when there's no traffic.

>
> **Implementation**
>
> Comments
>
>     * Code docstring coverage in production code is good, especially in `core.py`.
>
>     * I was unclear what was going on in `main` but one final look before submitting and it appears [you may be failing CI / CD](https://github.com/lucasimi/tda-mapper-python/actions/runs/15959545925).
>
>
> Recommend
>
>     * Unless there is a specific reason, you should consider removing 3.8 in `pyproject.toml` as it is EOL. _However, I have not opened an issue for this because I am unclear if maybe 3.8 was included past EOL for some specific reason._
>

Support for Python 3.8 has been dropped in the [pyproject.toml](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/pyproject.toml#L42) and [in the CI](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/.github/workflows/test-unit.yml#L16).

>     * The app itself (`app.py`, `streamlit_app.py`) may benefit from inline documentation. _However, I have not opened an issue to have discussion here if you feel strongly._
>

Inline comments have been added to the nicegui app. The streamlit app is deprecated, so I would like not to add inline documentation because it's not useful anymore at this point. The core functionalities have been migrated to the newer nicegui app. However, some minor features are still missing and need to be migrated. For these reasons I would like to keep the streamlit app in the main branch, as long as it's necessary to complete the migration, but also avoid any new addition to its code.

>
> Consider (optional)
>
>     * I would still suggest adding at least short descriptions to supporting structures with underscore prefix like to `Pullback` and `MapperAlgorithm`. However, these are not public members so this is optional. However, even just a description of their purpose would be helpful.
>

I agree, those details are actually important. Inline docstrings have been added [here](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/src/tdamapper/cover.py#L32).

>     * The tests are simple but coverage is broad. My only concern is that the the tests are not very varied and test data are not well described (like X in `test_unit_knn.py`).  Some comments here may help.
>

I've removed some old tests and refactored the majority of them using pytest's parametric testing decorators. This improved the variability of the tests, and also highlighted some minor issues that have now been solved. The majority of the tests accept a dataset parameter that ranges over some simple use cases where it's easy to reason about ([those are now documented](https://github.com/lucasimi/tda-mapper-python/blob/main/tests/test_utils.py)), and a random dataset that helps in keeping the test more varied. Some older test have been removed, for example `test_unit_knn.py`, since KNN is already tested both in `test_unit_cover.py` and in `test_unit_vptree.py`.

>     * You have type information in a number of docstrings but you don't use actual type hints. For  the places where this is available, this would help mypy users.
>

Type hints have been added throughout all the code.

>     * I know its fairly straightforward but you should consider adding a docstring to `EstimatorMixin` similar to `ParamsMixin`. In general, public members should have documentation even if just for other developers.

I agree. Proper docstrings have been added [here](https://github.com/lucasimi/tda-mapper-python/blob/b921c635302b3085efaabb7312bc4feb5bc3bba0/src/tdamapper/_common.py#L46).

Please, feel free to share your feedback. If you find anything missing or needing further expansion I'll be happy to adjust.
