### 2025-10-13 #222 @yeelauren
Hey everyone, thanks  @elliesch for agreeing to review, please see this comment here to get started 🪂
https://github.com/pyOpenSci/software-submission/issues/222#issuecomment-2956354434
